more to come about shrinking the ObjectStore
