+++ standalone-metastore/metastore-server/src/main/java/org/apache/hadoop/hive/metastore/ObjectStore.java
@@ -22,15 +22,10 @@ import static org.apache.hadoop.hive.metastore.Batchable.NO_BATCHING;
 import static org.apache.hadoop.hive.metastore.conf.MetastoreConf.ConfVars.COMPACTOR_USE_CUSTOM_POOL;
 import static org.apache.hadoop.hive.metastore.utils.MetaStoreUtils.getDefaultCatalog;
 import static org.apache.hadoop.hive.metastore.utils.StringUtils.normalizeIdentifier;
-import static org.apache.hadoop.hive.metastore.utils.StringUtils.normalizeIdentifiers;
 
 import java.io.IOException;
 import java.net.InetAddress;
-import java.sql.Connection;
-import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.sql.SQLIntegrityConstraintViolationException;
-import java.sql.Statement;
 import java.time.Instant;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
@@ -50,11 +45,8 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.TreeMap;
 import java.util.UUID;
-import java.util.concurrent.ThreadLocalRandom;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicReference;
-import java.util.function.Consumer;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
@@ -64,48 +56,29 @@ import javax.jdo.JDOObjectNotFoundException;
 import javax.jdo.PersistenceManager;
 import javax.jdo.Query;
 import javax.jdo.Transaction;
-import javax.jdo.datastore.JDOConnection;
 import javax.jdo.identity.IntIdentity;
 
 import org.apache.commons.collections4.CollectionUtils;
-import org.apache.commons.lang3.ArrayUtils;
 import org.apache.commons.lang3.StringUtils;
-import org.apache.commons.lang3.exception.ExceptionUtils;
-import org.apache.commons.lang3.tuple.Pair;
 import org.apache.hadoop.classification.InterfaceAudience;
 import org.apache.hadoop.classification.InterfaceStability;
 import org.apache.hadoop.conf.Configurable;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.hive.common.DatabaseName;
 import org.apache.hadoop.hive.common.StatsSetupConst;
-import org.apache.hadoop.hive.common.TableName;
-import org.apache.hadoop.hive.common.ValidReaderWriteIdList;
-import org.apache.hadoop.hive.common.ValidWriteIdList;
-import org.apache.hadoop.hive.metastore.directsql.DirectSqlDeleteStats;
 import org.apache.hadoop.hive.metastore.directsql.MetaStoreDirectSql;
-import org.apache.hadoop.hive.metastore.directsql.MetaStoreDirectSql.SqlFilterForPushdown;
-import org.apache.hadoop.hive.metastore.api.AggrStats;
-import org.apache.hadoop.hive.metastore.api.AllTableConstraintsRequest;
 import org.apache.hadoop.hive.metastore.api.AlreadyExistsException;
 import org.apache.hadoop.hive.metastore.api.Catalog;
-import org.apache.hadoop.hive.metastore.api.CheckConstraintsRequest;
-import org.apache.hadoop.hive.metastore.api.ColumnStatistics;
-import org.apache.hadoop.hive.metastore.api.ColumnStatisticsDesc;
-import org.apache.hadoop.hive.metastore.api.ColumnStatisticsObj;
 import org.apache.hadoop.hive.metastore.api.CreationMetadata;
 import org.apache.hadoop.hive.metastore.api.Database;
 import org.apache.hadoop.hive.metastore.api.AddPackageRequest;
-import org.apache.hadoop.hive.metastore.api.DefaultConstraintsRequest;
 import org.apache.hadoop.hive.metastore.api.DropPackageRequest;
 import org.apache.hadoop.hive.metastore.api.DatabaseType;
 import org.apache.hadoop.hive.metastore.api.DataConnector;
 import org.apache.hadoop.hive.metastore.api.FieldSchema;
-import org.apache.hadoop.hive.metastore.api.ForeignKeysRequest;
 import org.apache.hadoop.hive.metastore.api.Function;
 import org.apache.hadoop.hive.metastore.api.FunctionType;
 import org.apache.hadoop.hive.metastore.api.GetPackageRequest;
-import org.apache.hadoop.hive.metastore.api.GetPartitionsFilterSpec;
-import org.apache.hadoop.hive.metastore.api.GetProjectionsSpec;
 import org.apache.hadoop.hive.metastore.api.ISchema;
 import org.apache.hadoop.hive.metastore.api.ISchemaName;
 import org.apache.hadoop.hive.metastore.api.InvalidInputException;
@@ -115,14 +88,10 @@ import org.apache.hadoop.hive.metastore.api.ListPackageRequest;
 import org.apache.hadoop.hive.metastore.api.ListStoredProcedureRequest;
 import org.apache.hadoop.hive.metastore.api.MetaException;
 import org.apache.hadoop.hive.metastore.api.NoSuchObjectException;
-import org.apache.hadoop.hive.metastore.api.NotNullConstraintsRequest;
 import org.apache.hadoop.hive.metastore.api.Order;
 import org.apache.hadoop.hive.metastore.api.Package;
 import org.apache.hadoop.hive.metastore.api.Partition;
-import org.apache.hadoop.hive.metastore.api.PartitionFilterMode;
-import org.apache.hadoop.hive.metastore.api.PrimaryKeysRequest;
 import org.apache.hadoop.hive.metastore.api.PrincipalType;
-import org.apache.hadoop.hive.metastore.api.PrivilegeGrantInfo;
 import org.apache.hadoop.hive.metastore.api.QueryState;
 import org.apache.hadoop.hive.metastore.api.ResourceType;
 import org.apache.hadoop.hive.metastore.api.ResourceUri;
@@ -130,13 +99,6 @@ import org.apache.hadoop.hive.metastore.api.RuntimeStat;
 import org.apache.hadoop.hive.metastore.api.ReplicationMetricList;
 import org.apache.hadoop.hive.metastore.api.GetReplicationMetricsRequest;
 import org.apache.hadoop.hive.metastore.api.ReplicationMetrics;
-import org.apache.hadoop.hive.metastore.api.SQLAllTableConstraints;
-import org.apache.hadoop.hive.metastore.api.SQLCheckConstraint;
-import org.apache.hadoop.hive.metastore.api.SQLDefaultConstraint;
-import org.apache.hadoop.hive.metastore.api.SQLForeignKey;
-import org.apache.hadoop.hive.metastore.api.SQLNotNullConstraint;
-import org.apache.hadoop.hive.metastore.api.SQLPrimaryKey;
-import org.apache.hadoop.hive.metastore.api.SQLUniqueConstraint;
 import org.apache.hadoop.hive.metastore.api.ScheduledQuery;
 import org.apache.hadoop.hive.metastore.api.ScheduledQueryKey;
 import org.apache.hadoop.hive.metastore.api.ScheduledQueryMaintenanceRequest;
@@ -158,7 +120,6 @@ import org.apache.hadoop.hive.metastore.api.StoredProcedure;
 import org.apache.hadoop.hive.metastore.api.Table;
 import org.apache.hadoop.hive.metastore.api.TableParamsUpdate;
 import org.apache.hadoop.hive.metastore.api.Type;
-import org.apache.hadoop.hive.metastore.api.UniqueConstraintsRequest;
 import org.apache.hadoop.hive.metastore.api.WMFullResourcePlan;
 import org.apache.hadoop.hive.metastore.api.WMMapping;
 import org.apache.hadoop.hive.metastore.api.WMNullablePool;
@@ -178,7 +139,6 @@ import org.apache.hadoop.hive.metastore.metrics.Metrics;
 import org.apache.hadoop.hive.metastore.metrics.MetricsConstants;
 import org.apache.hadoop.hive.metastore.model.MCatalog;
 import org.apache.hadoop.hive.metastore.model.MColumnDescriptor;
-import org.apache.hadoop.hive.metastore.model.MConstraint;
 import org.apache.hadoop.hive.metastore.model.MCreationMetadata;
 import org.apache.hadoop.hive.metastore.model.MDBPrivilege;
 import org.apache.hadoop.hive.metastore.model.MDataConnector;
@@ -194,7 +154,6 @@ import org.apache.hadoop.hive.metastore.model.MMetastoreDBProperties;
 import org.apache.hadoop.hive.metastore.model.MOrder;
 import org.apache.hadoop.hive.metastore.model.MPackage;
 import org.apache.hadoop.hive.metastore.model.MPartition;
-import org.apache.hadoop.hive.metastore.model.MPartitionColumnStatistics;
 import org.apache.hadoop.hive.metastore.model.MPartitionEvent;
 import org.apache.hadoop.hive.metastore.model.MResourceUri;
 import org.apache.hadoop.hive.metastore.model.MRuntimeStat;
@@ -206,8 +165,6 @@ import org.apache.hadoop.hive.metastore.model.MStorageDescriptor;
 import org.apache.hadoop.hive.metastore.model.MStoredProc;
 import org.apache.hadoop.hive.metastore.model.MStringList;
 import org.apache.hadoop.hive.metastore.model.MTable;
-import org.apache.hadoop.hive.metastore.model.MTableColumnStatistics;
-import org.apache.hadoop.hive.metastore.model.MTablePrivilege;
 import org.apache.hadoop.hive.metastore.model.MType;
 import org.apache.hadoop.hive.metastore.model.MVersionTable;
 import org.apache.hadoop.hive.metastore.model.MWMMapping;
@@ -228,7 +185,6 @@ import org.apache.hadoop.hive.metastore.txn.TxnUtils;
 import org.apache.hadoop.hive.metastore.utils.JavaUtils;
 import org.apache.hadoop.hive.metastore.utils.MetaStoreServerUtils;
 import org.apache.hadoop.hive.metastore.utils.MetaStoreUtils;
-import org.apache.hadoop.hive.metastore.utils.RetryingExecutor;
 import org.datanucleus.ExecutionContext;
 import org.datanucleus.api.jdo.JDOPersistenceManager;
 import org.datanucleus.api.jdo.JDOTransaction;
@@ -244,12 +200,9 @@ import com.cronutils.model.definition.CronDefinitionBuilder;
 import com.cronutils.model.time.ExecutionTime;
 import com.cronutils.parser.CronParser;
 import com.google.common.annotations.VisibleForTesting;
-import com.google.common.base.Preconditions;
 import com.google.common.base.Strings;
 import com.google.common.collect.Lists;
-import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
-
 
 /**
  * This class is the interface between the application logic and the database
@@ -1321,202 +1274,6 @@ public class ObjectStore implements RawStore, Configurable {
     return success;
   }
 
-  @Override
-  public SQLAllTableConstraints createTableWithConstraints(Table tbl, SQLAllTableConstraints constraints)
-      throws InvalidObjectException, MetaException {
-    boolean success = false;
-    try {
-      openTransaction();
-      createTable(tbl);
-      // Add constraints.
-      // We need not do a deep retrieval of the Table Column Descriptor while persisting the
-      // constraints since this transaction involving create table is not yet committed.
-      if (CollectionUtils.isNotEmpty(constraints.getForeignKeys())) {
-        constraints.setForeignKeys(addForeignKeys(constraints.getForeignKeys(), false, constraints.getPrimaryKeys(),
-            constraints.getUniqueConstraints()));
-      }
-      if (CollectionUtils.isNotEmpty(constraints.getPrimaryKeys())) {
-        constraints.setPrimaryKeys(addPrimaryKeys(constraints.getPrimaryKeys(), false));
-      }
-      if (CollectionUtils.isNotEmpty(constraints.getUniqueConstraints())) {
-        constraints.setUniqueConstraints(addUniqueConstraints(constraints.getUniqueConstraints(), false));
-      }
-      if (CollectionUtils.isNotEmpty(constraints.getNotNullConstraints())) {
-        constraints.setNotNullConstraints(addNotNullConstraints(constraints.getNotNullConstraints(), false));
-      }
-      if (CollectionUtils.isNotEmpty(constraints.getDefaultConstraints())) {
-        constraints.setDefaultConstraints(addDefaultConstraints(constraints.getDefaultConstraints(), false));
-      }
-      if (CollectionUtils.isNotEmpty(constraints.getCheckConstraints())) {
-        constraints.setCheckConstraints(addCheckConstraints(constraints.getCheckConstraints(), false));
-      }
-      success = commitTransaction();
-      return constraints;
-    } finally {
-      rollbackAndCleanup(success, null);
-    }
-  }
-
-  /**
-   * Convert PrivilegeGrantInfo from privMap to MTablePrivilege, and add all of
-   * them to the toPersistPrivObjs. These privilege objects will be persisted as
-   * part of createTable.
-   */
-  public static void putPersistentPrivObjects(MTable mtbl, List<Object> toPersistPrivObjs,
-      int now, Map<String, List<PrivilegeGrantInfo>> privMap, PrincipalType type, String authorizer) {
-    if (privMap != null) {
-      for (Map.Entry<String, List<PrivilegeGrantInfo>> entry : privMap
-          .entrySet()) {
-        String principalName = entry.getKey();
-        List<PrivilegeGrantInfo> privs = entry.getValue();
-        for (int i = 0; i < privs.size(); i++) {
-          PrivilegeGrantInfo priv = privs.get(i);
-          if (priv == null) {
-            continue;
-          }
-          MTablePrivilege mTblSec = new MTablePrivilege(
-              principalName, type.toString(), mtbl, priv.getPrivilege(),
-              now, priv.getGrantor(), priv.getGrantorType().toString(), priv
-                  .isGrantOption(), authorizer);
-          toPersistPrivObjs.add(mTblSec);
-        }
-      }
-    }
-  }
-
-  private List<MConstraint> listAllTableConstraintsWithOptionalConstraintName(
-      String catName, String dbName, String tableName, String constraintname) {
-    catName = normalizeIdentifier(catName);
-    dbName = normalizeIdentifier(dbName);
-    tableName = normalizeIdentifier(tableName);
-    constraintname = constraintname!=null?normalizeIdentifier(constraintname):null;
-    List<MConstraint> mConstraints = null;
-    List<String> constraintNames = new ArrayList<>();
-
-    try (QueryWrapper queryForConstraintName = new QueryWrapper(pm.newQuery("select constraintName from org.apache.hadoop.hive.metastore.model.MConstraint  where "
-        + "((parentTable.tableName == ptblname && parentTable.database.name == pdbname && " +
-        "parentTable.database.catalogName == pcatname) || "
-        + "(childTable != null && childTable.tableName == ctblname &&" +
-        "childTable.database.name == cdbname && childTable.database.catalogName == ccatname)) " +
-        (constraintname != null ? " && constraintName == constraintname" : "")));
-        QueryWrapper queryForMConstraint = new QueryWrapper(pm.newQuery(MConstraint.class))) {
-
-      queryForConstraintName.declareParameters("java.lang.String ptblname, java.lang.String pdbname,"
-          + "java.lang.String pcatname, java.lang.String ctblname, java.lang.String cdbname," +
-          "java.lang.String ccatname" +
-        (constraintname != null ? ", java.lang.String constraintname" : ""));
-      Collection<?> constraintNamesColl =
-        constraintname != null ?
-          ((Collection<?>) queryForConstraintName.
-            executeWithArray(tableName, dbName, catName, tableName, dbName, catName, constraintname)):
-          ((Collection<?>) queryForConstraintName.
-            executeWithArray(tableName, dbName, catName, tableName, dbName, catName));
-      for (Iterator<?> i = constraintNamesColl.iterator(); i.hasNext();) {
-        String currName = (String) i.next();
-        constraintNames.add(currName);
-      }
-
-      queryForMConstraint.setFilter("param.contains(constraintName)");
-      queryForMConstraint.declareParameters("java.util.Collection param");
-      Collection<?> constraints = (Collection<?>)queryForMConstraint.execute(constraintNames);
-      mConstraints = new ArrayList<>();
-      for (Iterator<?> i = constraints.iterator(); i.hasNext();) {
-        MConstraint currConstraint = (MConstraint) i.next();
-        mConstraints.add(currConstraint);
-      }
-    }
-    return mConstraints;
-  }
-
-  @Override
-  public List<TableName> getTableNamesWithStats() throws MetaException, NoSuchObjectException {
-    return new GetListHelper<TableName>(null, null, null, true, false) {
-      @Override
-      protected List<TableName> getSqlResult(
-          GetHelper<List<TableName>> ctx) throws MetaException {
-        return directSql.getTableNamesWithStats();
-      }
-
-      @Override
-      protected List<TableName> getJdoResult(
-          GetHelper<List<TableName>> ctx) throws MetaException {
-        throw new UnsupportedOperationException("UnsupportedOperationException"); // TODO: implement?
-      }
-    }.run(false);
-  }
-
-  @Override
-  public Map<String, List<String>> getPartitionColsWithStats(String catName, String dbName, String tableName)
-      throws MetaException, NoSuchObjectException {
-    return new GetHelper<Map<String, List<String>>>(catName, dbName, null, true, false) {
-      @Override
-      protected Map<String, List<String>> getSqlResult(
-          GetHelper<Map<String, List<String>>> ctx) throws MetaException {
-        try {
-          return directSql.getColAndPartNamesWithStats(catName, dbName, tableName);
-        } catch (Throwable ex) {
-          LOG.error("DirectSQL failed", ex);
-          throw new MetaException(ex.getMessage());
-        }
-      }
-
-      @Override
-      protected Map<String, List<String>> getJdoResult(
-          GetHelper<Map<String, List<String>>> ctx) throws MetaException {
-        throw new UnsupportedOperationException("UnsupportedOperationException"); // TODO: implement?
-      }
-
-      @Override
-      protected String describeResult() {
-        return results.size() + " partitions";
-      }
-    }.run(false);
-  }
-
-  @Override
-  public List<TableName> getAllTableNamesForStats() throws MetaException, NoSuchObjectException {
-    return new GetListHelper<TableName>(null, null, null, true, false) {
-      @Override
-      protected List<TableName> getSqlResult(
-          GetHelper<List<TableName>> ctx) throws MetaException {
-        return directSql.getAllTableNamesForStats();
-      }
-
-      @Override
-      protected List<TableName> getJdoResult(
-          GetHelper<List<TableName>> ctx) throws MetaException {
-        boolean commited = false;
-        Query query = null;
-        List<TableName> result = new ArrayList<>();
-        openTransaction();
-        try {
-          String paramStr = "", whereStr = "";
-          for (int i = 0; i < MetaStoreDirectSql.STATS_TABLE_TYPES.length; ++i) {
-            if (i != 0) {
-              paramStr += ", ";
-              whereStr += "||";
-            }
-            paramStr += "java.lang.String tt" + i;
-            whereStr += " tableType == tt" + i;
-          }
-          query = pm.newQuery(MTable.class, whereStr);
-          query.declareParameters(paramStr);
-          Collection<MTable> tbls = (Collection<MTable>) query.executeWithArray(
-              query, MetaStoreDirectSql.STATS_TABLE_TYPES);
-          pm.retrieveAll(tbls);
-          for (MTable tbl : tbls) {
-            result.add(new TableName(
-                tbl.getDatabase().getCatalogName(), tbl.getDatabase().getName(), tbl.getTableName()));
-          }
-          commited = commitTransaction();
-        } finally {
-          rollbackAndCleanup(commited, query);
-        }
-        return result;
-      }
-    }.run(false);
-  }
-
   public static StringBuilder appendPatternCondition(StringBuilder builder,
       String fieldName, String elements, List<String> parameters) {
       elements = normalizeIdentifier(elements);
@@ -1552,59 +1309,6 @@ public class ObjectStore implements RawStore, Configurable {
     }
     builder.append(" )");
     return builder;
-  }
-
-  class AttachedMTableInfo {
-    MTable mtbl;
-    MColumnDescriptor mcd;
-
-    public AttachedMTableInfo() {}
-
-    public AttachedMTableInfo(MTable mtbl, MColumnDescriptor mcd) {
-      this.mtbl = mtbl;
-      this.mcd = mcd;
-    }
-  }
-
-  private AttachedMTableInfo getMTable(String catName, String db, String table,
-                                       boolean retrieveCD) {
-    AttachedMTableInfo nmtbl = new AttachedMTableInfo();
-    MTable mtbl = null;
-    boolean commited = false;
-    Query query = null;
-    try {
-      openTransaction();
-      catName = normalizeIdentifier(Optional.ofNullable(catName).orElse(getDefaultCatalog(conf)));
-      db = normalizeIdentifier(db);
-      table = normalizeIdentifier(table);
-      query = pm.newQuery(MTable.class,
-          "tableName == table && database.name == db && database.catalogName == catname");
-      query.declareParameters(
-          "java.lang.String table, java.lang.String db, java.lang.String catname");
-      query.setUnique(true);
-      if (LOG.isDebugEnabled()) {
-        LOG.debug("Executing getMTable for {}",
-            TableName.getQualified(catName, db, table));
-      }
-      mtbl = (MTable) query.execute(table, db, catName);
-      pm.retrieve(mtbl);
-      // Retrieving CD can be expensive and unnecessary, so do it only when required.
-      if (mtbl != null && retrieveCD) {
-        pm.retrieve(mtbl.getSd());
-        pm.retrieveAll(mtbl.getSd().getCD());
-        nmtbl.mcd = mtbl.getSd().getCD();
-      }
-      commited = commitTransaction();
-    } finally {
-      rollbackAndCleanup(commited, query);
-    }
-    nmtbl.mtbl = mtbl;
-    return nmtbl;
-  }
-
-  private MTable getMTable(String catName, String db, String table) {
-    AttachedMTableInfo nmtbl = getMTable(catName, db, table, false);
-    return nmtbl.mtbl;
   }
 
   /** Makes shallow copy of a list to avoid DataNucleus mucking with our objects. */
@@ -2132,49 +1836,6 @@ public class ObjectStore implements RawStore, Configurable {
     return parts;
   }
 
-  public static Pair<Query, Map<String, String>> getPartQueryWithParams(
-      PersistenceManager pm,
-      String catName, String dbName, String tblName,
-      List<String> partNames) {
-    Query query = pm.newQuery();
-    Map<String, String> params = new HashMap<>();
-    String filterStr = getJDOFilterStrForPartitionNames(catName, dbName, tblName, partNames, params);
-    query.setFilter(filterStr);
-    LOG.debug(" JDOQL filter is {}", filterStr);
-    query.declareParameters(makeParameterDeclarationString(params));
-    return Pair.of(query, params);
-  }
-
-  public static String getJDOFilterStrForPartitionNames(String catName, String dbName, String tblName,
-      List<String> partNames, Map params) {
-    StringBuilder sb = new StringBuilder(
-        "table.tableName == t1 && table.database.name == t2 &&" + " table.database.catalogName == t3 && (");
-    params.put("t1", normalizeIdentifier(tblName));
-    params.put("t2", normalizeIdentifier(dbName));
-    params.put("t3", normalizeIdentifier(catName));
-    int n = 0;
-    for (Iterator<String> itr = partNames.iterator(); itr.hasNext(); ) {
-      String pn = "p" + n;
-      n++;
-      String part = itr.next();
-      params.put(pn, part);
-      sb.append("partitionName == ").append(pn);
-      sb.append(" || ");
-    }
-    sb.setLength(sb.length() - 4); // remove the last " || "
-    sb.append(')');
-    return sb.toString();
-  }
-
-  public static String makeParameterDeclarationString(Map<String, String> params) {
-    //Create the parameter declaration string
-    StringBuilder paramDecl = new StringBuilder();
-    for (String key : params.keySet()) {
-      paramDecl.append(", java.lang.String ").append(key);
-    }
-    return paramDecl.toString();
-  }
-
   /** Helper class for getting stuff w/transaction, direct SQL, perf logging, etc. */
   @VisibleForTesting
   public abstract class GetHelper<T> {
@@ -2407,18 +2068,6 @@ public class ObjectStore implements RawStore, Configurable {
     }
   }
 
-  private abstract class GetStatHelper extends GetHelper<ColumnStatistics> {
-    public GetStatHelper(String catalogName, String dbName, String tblName, boolean allowSql,
-                         boolean allowJdo, String writeIdList) throws MetaException {
-      super(catalogName, dbName, tblName, allowSql, allowJdo);
-    }
-
-    @Override
-    protected String describeResult() {
-      return "statistics for " + (results == null ? 0 : results.getStatsObjSize()) + " columns";
-    }
-  }
-
   private Table ensureGetTable(String catName, String dbName, String tblName)
       throws NoSuchObjectException, MetaException {
     return convertToTable(ensureGetMTable(catName, dbName, tblName), conf);
@@ -2453,71 +2102,6 @@ public class ObjectStore implements RawStore, Configurable {
             "providing the transactional write state for verification (new write ID " +
             writeId + ", valid write IDs " + validWriteIds + "; current state " + oldVal + "; new" +
             " state " + newVal;
-  }
-
-  private static MFieldSchema getColumnFromTableColumns(List<MFieldSchema> cols, String col) {
-    if (cols == null) {
-      return null;
-    }
-    for (MFieldSchema mfs : cols) {
-      if (mfs.getName().equalsIgnoreCase(col)) {
-        return mfs;
-      }
-    }
-    return null;
-  }
-
-  private static int getColumnIndexFromTableColumns(List<MFieldSchema> cols, String col) {
-    if (cols == null) {
-      return -1;
-    }
-    for (int i = 0; i < cols.size(); i++) {
-      MFieldSchema mfs = cols.get(i);
-      if (mfs.getName().equalsIgnoreCase(col)) {
-        return i;
-      }
-    }
-    return -1;
-  }
-
-  private  boolean constraintNameAlreadyExists(MTable table, String constraintName) {
-    boolean commited = false;
-    Query<MConstraint> constraintExistsQuery = null;
-    String constraintNameIfExists = null;
-    try {
-      openTransaction();
-      constraintName = normalizeIdentifier(constraintName);
-      constraintExistsQuery = pm.newQuery(MConstraint.class,
-          "parentTable == parentTableP && constraintName == constraintNameP");
-      constraintExistsQuery.declareParameters("MTable parentTableP, java.lang.String constraintNameP");
-      constraintExistsQuery.setUnique(true);
-      constraintExistsQuery.setResult("constraintName");
-      constraintNameIfExists = (String) constraintExistsQuery.executeWithArray(table, constraintName);
-      commited = commitTransaction();
-    } finally {
-      rollbackAndCleanup(commited, constraintExistsQuery);
-    }
-    return constraintNameIfExists != null && !constraintNameIfExists.isEmpty();
-  }
-
-  private String generateConstraintName(MTable table, String... parameters) throws MetaException {
-    int hashcode = ArrayUtils.toString(parameters).hashCode() & 0xfffffff;
-    int counter = 0;
-    final int MAX_RETRIES = 10;
-    while (counter < MAX_RETRIES) {
-      String currName = (parameters.length == 0 ? "constraint_" : parameters[parameters.length-1]) +
-        "_" + hashcode + "_" + System.currentTimeMillis() + "_" + (counter++);
-      if (!constraintNameAlreadyExists(table, currName)) {
-        return currName;
-      }
-    }
-    throw new MetaException("Error while trying to generate the constraint name for " + ArrayUtils.toString(parameters));
-  }
-
-  @Override
-  public List<SQLForeignKey> addForeignKeys(
-    List<SQLForeignKey> fks) throws InvalidObjectException, MetaException {
-   return addForeignKeys(fks, true, null, null);
   }
 
   @Override
@@ -2771,1642 +2355,6 @@ public class ObjectStore implements RawStore, Configurable {
       rollbackAndCleanup(success, query);
     }
     return results;
-  }
-
-  //TODO: clean up this method
-  private List<SQLForeignKey> addForeignKeys(List<SQLForeignKey> foreignKeys, boolean retrieveCD,
-      List<SQLPrimaryKey> primaryKeys, List<SQLUniqueConstraint> uniqueConstraints)
-          throws InvalidObjectException, MetaException {
-    if (CollectionUtils.isNotEmpty(foreignKeys)) {
-      List<MConstraint> mpkfks = new ArrayList<>();
-      String currentConstraintName = null;
-      String catName = null;
-      // We start iterating through the foreign keys. This list might contain more than a single
-      // foreign key, and each foreign key might contain multiple columns. The outer loop retrieves
-      // the information that is common for a single key (table information) while the inner loop
-      // checks / adds information about each column.
-      for (int i = 0; i < foreignKeys.size(); i++) {
-        if (catName == null) {
-          catName = normalizeIdentifier(foreignKeys.get(i).isSetCatName() ? foreignKeys.get(i).getCatName() :
-              getDefaultCatalog(conf));
-        } else {
-          String tmpCatName = normalizeIdentifier(foreignKeys.get(i).isSetCatName() ?
-              foreignKeys.get(i).getCatName() : getDefaultCatalog(conf));
-          if (!catName.equals(tmpCatName)) {
-            throw new InvalidObjectException("Foreign keys cannot span catalogs");
-          }
-        }
-        final String fkTableDB = normalizeIdentifier(foreignKeys.get(i).getFktable_db());
-        final String fkTableName = normalizeIdentifier(foreignKeys.get(i).getFktable_name());
-        // If retrieveCD is false, we do not need to do a deep retrieval of the Table Column Descriptor.
-        // For instance, this is the case when we are creating the table.
-        final AttachedMTableInfo nChildTable = getMTable(catName, fkTableDB, fkTableName, retrieveCD);
-        final MTable childTable = nChildTable.mtbl;
-        if (childTable == null) {
-          throw new InvalidObjectException("Child table not found: " + fkTableName);
-        }
-        MColumnDescriptor childCD = retrieveCD ? nChildTable.mcd : childTable.getSd().getCD();
-        final List<MFieldSchema> childCols = childCD == null || childCD.getCols() == null ?
-            new ArrayList<>() : new ArrayList<>(childCD.getCols());
-        if (childTable.getPartitionKeys() != null) {
-          childCols.addAll(childTable.getPartitionKeys());
-        }
-
-        final String pkTableDB = normalizeIdentifier(foreignKeys.get(i).getPktable_db());
-        final String pkTableName = normalizeIdentifier(foreignKeys.get(i).getPktable_name());
-        // For primary keys, we retrieve the column descriptors if retrieveCD is true (which means
-        // it is an alter table statement) or if it is a create table statement but we are
-        // referencing another table instead of self for the primary key.
-        final AttachedMTableInfo nParentTable;
-        final MTable parentTable;
-        MColumnDescriptor parentCD;
-        final List<MFieldSchema> parentCols;
-        final List<SQLPrimaryKey> existingTablePrimaryKeys;
-        final List<SQLUniqueConstraint> existingTableUniqueConstraints;
-        final boolean sameTable = fkTableDB.equals(pkTableDB) && fkTableName.equals(pkTableName);
-        if (sameTable) {
-          nParentTable = nChildTable;
-          parentTable = childTable;
-          parentCD = childCD;
-          parentCols = childCols;
-          existingTablePrimaryKeys = primaryKeys;
-          existingTableUniqueConstraints = uniqueConstraints;
-        } else {
-          nParentTable = getMTable(catName, pkTableDB, pkTableName, true);
-          parentTable = nParentTable.mtbl;
-          if (parentTable == null) {
-            throw new InvalidObjectException("Parent table not found: " + pkTableName);
-          }
-          parentCD = nParentTable.mcd;
-          parentCols = parentCD == null || parentCD.getCols() == null ?
-              new ArrayList<>() : new ArrayList<>(parentCD.getCols());
-          if (parentTable.getPartitionKeys() != null) {
-            parentCols.addAll(parentTable.getPartitionKeys());
-          }
-          PrimaryKeysRequest primaryKeysRequest = new PrimaryKeysRequest(pkTableDB, pkTableName);
-          primaryKeysRequest.setCatName(catName);
-          existingTablePrimaryKeys = getPrimaryKeys(primaryKeysRequest);
-          existingTableUniqueConstraints =
-              getUniqueConstraints(new UniqueConstraintsRequest(catName, pkTableDB, pkTableName));
-        }
-
-        // Here we build an aux structure that is used to verify that the foreign key that is declared
-        // is actually referencing a valid primary key or unique key. We also check that the types of
-        // the columns correspond.
-        if (existingTablePrimaryKeys.isEmpty() && existingTableUniqueConstraints.isEmpty()) {
-          throw new MetaException(
-              "Trying to define foreign key but there are no primary keys or unique keys for referenced table");
-        }
-        final Set<String> validPKsOrUnique = generateValidPKsOrUniqueSignatures(parentCols,
-            existingTablePrimaryKeys, existingTableUniqueConstraints);
-
-        StringBuilder fkSignature = new StringBuilder();
-        StringBuilder referencedKSignature = new StringBuilder();
-        for (; i < foreignKeys.size(); i++) {
-          SQLForeignKey foreignKey = foreignKeys.get(i);
-          final String fkColumnName = normalizeIdentifier(foreignKey.getFkcolumn_name());
-          int childIntegerIndex = getColumnIndexFromTableColumns(childCD.getCols(), fkColumnName);
-          if (childIntegerIndex == -1) {
-            if (childTable.getPartitionKeys() != null) {
-              childCD = null;
-              childIntegerIndex = getColumnIndexFromTableColumns(childTable.getPartitionKeys(), fkColumnName);
-            }
-            if (childIntegerIndex == -1) {
-              throw new InvalidObjectException("Child column not found: " + fkColumnName);
-            }
-          }
-
-          final String pkColumnName = normalizeIdentifier(foreignKey.getPkcolumn_name());
-          int parentIntegerIndex = getColumnIndexFromTableColumns(parentCD.getCols(), pkColumnName);
-          if (parentIntegerIndex == -1) {
-            if (parentTable.getPartitionKeys() != null) {
-              parentCD = null;
-              parentIntegerIndex = getColumnIndexFromTableColumns(parentTable.getPartitionKeys(), pkColumnName);
-            }
-            if (parentIntegerIndex == -1) {
-              throw new InvalidObjectException("Parent column not found: " + pkColumnName);
-            }
-          }
-
-          if (foreignKey.getFk_name() == null) {
-            // When there is no explicit foreign key name associated with the constraint and the key is composite,
-            // we expect the foreign keys to be send in order in the input list.
-            // Otherwise, the below code will break.
-            // If this is the first column of the FK constraint, generate the foreign key name
-            // NB: The below code can result in race condition where duplicate names can be generated (in theory).
-            // However, this scenario can be ignored for practical purposes because of
-            // the uniqueness of the generated constraint name.
-            if (foreignKey.getKey_seq() == 1) {
-              currentConstraintName = generateConstraintName(parentTable, fkTableDB, fkTableName, pkTableDB,
-                  pkTableName, pkColumnName, fkColumnName, "fk");
-            }
-          } else {
-            currentConstraintName = normalizeIdentifier(foreignKey.getFk_name());
-            if (constraintNameAlreadyExists(parentTable, currentConstraintName)) {
-              String fqConstraintName = String.format("%s.%s.%s", parentTable.getDatabase().getName(),
-                  parentTable.getTableName(), currentConstraintName);
-              throw new InvalidObjectException("Constraint name already exists: " + fqConstraintName);
-            }
-          }
-          // Update Column, keys, table, database, catalog name
-          foreignKey.setFk_name(currentConstraintName);
-          foreignKey.setCatName(catName);
-          foreignKey.setFktable_db(fkTableDB);
-          foreignKey.setFktable_name(fkTableName);
-          foreignKey.setPktable_db(pkTableDB);
-          foreignKey.setPktable_name(pkTableName);
-          foreignKey.setFkcolumn_name(fkColumnName);
-          foreignKey.setPkcolumn_name(pkColumnName);
-
-          Integer updateRule = foreignKey.getUpdate_rule();
-          Integer deleteRule = foreignKey.getDelete_rule();
-          int enableValidateRely = (foreignKey.isEnable_cstr() ? 4 : 0) +
-                  (foreignKey.isValidate_cstr() ? 2 : 0) + (foreignKey.isRely_cstr() ? 1 : 0);
-
-          MConstraint mpkfk = new MConstraint(
-              currentConstraintName,
-              foreignKey.getKey_seq(),
-              MConstraint.FOREIGN_KEY_CONSTRAINT,
-              deleteRule,
-              updateRule,
-              enableValidateRely,
-              parentTable,
-              childTable,
-              parentCD,
-              childCD,
-              childIntegerIndex,
-              parentIntegerIndex
-          );
-          mpkfks.add(mpkfk);
-
-          final String fkColType = getColumnFromTableColumns(childCols, fkColumnName).getType();
-          fkSignature.append(
-              generateColNameTypeSignature(fkColumnName, fkColType));
-          referencedKSignature.append(
-              generateColNameTypeSignature(pkColumnName, fkColType));
-
-          if (i + 1 < foreignKeys.size() && foreignKeys.get(i + 1).getKey_seq() == 1) {
-            // Next one is a new key, we bail out from the inner loop
-            break;
-          }
-        }
-        String referenced = referencedKSignature.toString();
-        if (!validPKsOrUnique.contains(referenced)) {
-          throw new MetaException(
-              "Foreign key references " + referenced + " but no corresponding "
-              + "primary key or unique key exists. Possible keys: " + validPKsOrUnique);
-        }
-        if (sameTable && fkSignature.toString().equals(referenced)) {
-          throw new MetaException(
-              "Cannot be both foreign key and primary/unique key on same table: " + referenced);
-        }
-        fkSignature = new StringBuilder();
-        referencedKSignature = new StringBuilder();
-      }
-      pm.makePersistentAll(mpkfks);
-
-    }
-    return foreignKeys;
-  }
-
-  private static Set<String> generateValidPKsOrUniqueSignatures(List<MFieldSchema> tableCols,
-      List<SQLPrimaryKey> refTablePrimaryKeys, List<SQLUniqueConstraint> refTableUniqueConstraints) {
-    final Set<String> validPKsOrUnique = new HashSet<>();
-    if (!refTablePrimaryKeys.isEmpty()) {
-      refTablePrimaryKeys.sort((o1, o2) -> {
-        int keyNameComp = o1.getPk_name().compareTo(o2.getPk_name());
-        if (keyNameComp == 0) {
-          return Integer.compare(o1.getKey_seq(), o2.getKey_seq());
-        }
-        return keyNameComp;
-      });
-      StringBuilder pkSignature = new StringBuilder();
-      for (SQLPrimaryKey pk : refTablePrimaryKeys) {
-        pkSignature.append(
-            generateColNameTypeSignature(
-                pk.getColumn_name(), getColumnFromTableColumns(tableCols, pk.getColumn_name()).getType()));
-      }
-      validPKsOrUnique.add(pkSignature.toString());
-    }
-    if (!refTableUniqueConstraints.isEmpty()) {
-      refTableUniqueConstraints.sort((o1, o2) -> {
-        int keyNameComp = o1.getUk_name().compareTo(o2.getUk_name());
-        if (keyNameComp == 0) {
-          return Integer.compare(o1.getKey_seq(), o2.getKey_seq());
-        }
-        return keyNameComp;
-      });
-      StringBuilder ukSignature = new StringBuilder();
-      for (int j = 0; j < refTableUniqueConstraints.size(); j++) {
-        SQLUniqueConstraint uk = refTableUniqueConstraints.get(j);
-        ukSignature.append(
-            generateColNameTypeSignature(
-                uk.getColumn_name(), getColumnFromTableColumns(tableCols, uk.getColumn_name()).getType()));
-        if (j + 1 < refTableUniqueConstraints.size()) {
-          if (!refTableUniqueConstraints.get(j + 1).getUk_name().equals(
-                  refTableUniqueConstraints.get(j).getUk_name())) {
-            validPKsOrUnique.add(ukSignature.toString());
-            ukSignature = new StringBuilder();
-          }
-        } else {
-          validPKsOrUnique.add(ukSignature.toString());
-        }
-      }
-    }
-    return validPKsOrUnique;
-  }
-
-  private static String generateColNameTypeSignature(String colName, String colType) {
-    return colName + ":" + colType + ";";
-  }
-
-  @Override
-  public List<SQLPrimaryKey> addPrimaryKeys(List<SQLPrimaryKey> pks) throws InvalidObjectException,
-    MetaException {
-    return addPrimaryKeys(pks, true);
-  }
-
-  private List<SQLPrimaryKey> addPrimaryKeys(List<SQLPrimaryKey> pks, boolean retrieveCD) throws InvalidObjectException,
-    MetaException {
-    List<MConstraint> mpks = new ArrayList<>();
-    String constraintName = null;
-
-    for (SQLPrimaryKey pk : pks) {
-      final String catName = normalizeIdentifier(pk.getCatName());
-      final String tableDB = normalizeIdentifier(pk.getTable_db());
-      final String tableName = normalizeIdentifier(pk.getTable_name());
-      final String columnName = normalizeIdentifier(pk.getColumn_name());
-
-      // If retrieveCD is false, we do not need to do a deep retrieval of the Table Column Descriptor.
-      // For instance, this is the case when we are creating the table.
-      AttachedMTableInfo nParentTable = getMTable(catName, tableDB, tableName, retrieveCD);
-      MTable parentTable = nParentTable.mtbl;
-      if (parentTable == null) {
-        throw new InvalidObjectException("Parent table not found: " + tableName);
-      }
-
-      MColumnDescriptor parentCD = retrieveCD ? nParentTable.mcd : parentTable.getSd().getCD();
-      int parentIntegerIndex = getColumnIndexFromTableColumns(parentCD == null ? null : parentCD.getCols(), columnName);
-      if (parentIntegerIndex == -1) {
-        if (parentTable.getPartitionKeys() != null) {
-          parentCD = null;
-          parentIntegerIndex = getColumnIndexFromTableColumns(parentTable.getPartitionKeys(), columnName);
-        }
-        if (parentIntegerIndex == -1) {
-          throw new InvalidObjectException("Parent column not found: " + columnName);
-        }
-      }
-      if (getPrimaryKeyConstraintName(parentTable.getDatabase().getCatalogName(),
-          parentTable.getDatabase().getName(), parentTable.getTableName()) != null) {
-        throw new MetaException(" Primary key already exists for: " +
-            TableName.getQualified(catName, tableDB, tableName));
-      }
-      if (pk.getPk_name() == null) {
-        if (pk.getKey_seq() == 1) {
-          constraintName = generateConstraintName(parentTable, tableDB, tableName, columnName, "pk");
-        }
-      } else {
-        constraintName = normalizeIdentifier(pk.getPk_name());
-        if (constraintNameAlreadyExists(parentTable, constraintName)) {
-          String fqConstraintName = String.format("%s.%s.%s", parentTable.getDatabase().getName(),
-              parentTable.getTableName(), constraintName);
-          throw new InvalidObjectException("Constraint name already exists: " + fqConstraintName);
-        }
-      }
-
-      int enableValidateRely = (pk.isEnable_cstr() ? 4 : 0) +
-      (pk.isValidate_cstr() ? 2 : 0) + (pk.isRely_cstr() ? 1 : 0);
-      MConstraint mpk = new MConstraint(
-          constraintName,
-          pk.getKey_seq(),
-          MConstraint.PRIMARY_KEY_CONSTRAINT,
-          null,
-          null,
-          enableValidateRely,
-          parentTable,
-          null,
-          parentCD,
-          null,
-          null,
-        parentIntegerIndex);
-      mpks.add(mpk);
-
-      // Add normalized identifier back to result
-      pk.setCatName(catName);
-      pk.setTable_db(tableDB);
-      pk.setTable_name(tableName);
-      pk.setColumn_name(columnName);
-      pk.setPk_name(constraintName);
-    }
-    pm.makePersistentAll(mpks);
-    return pks;
-  }
-
-  @Override
-  public List<SQLUniqueConstraint> addUniqueConstraints(List<SQLUniqueConstraint> uks)
-          throws InvalidObjectException, MetaException {
-    return addUniqueConstraints(uks, true);
-  }
-
-  private List<SQLUniqueConstraint> addUniqueConstraints(List<SQLUniqueConstraint> uks, boolean retrieveCD)
-          throws InvalidObjectException, MetaException {
-
-    List<MConstraint> cstrs = new ArrayList<>();
-    String constraintName = null;
-
-    for (SQLUniqueConstraint uk : uks) {
-      final String catName = normalizeIdentifier(uk.getCatName());
-      final String tableDB = normalizeIdentifier(uk.getTable_db());
-      final String tableName = normalizeIdentifier(uk.getTable_name());
-      final String columnName = normalizeIdentifier(uk.getColumn_name());
-
-      // If retrieveCD is false, we do not need to do a deep retrieval of the Table Column Descriptor.
-      // For instance, this is the case when we are creating the table.
-      AttachedMTableInfo nParentTable = getMTable(catName, tableDB, tableName, retrieveCD);
-      MTable parentTable = nParentTable.mtbl;
-      if (parentTable == null) {
-        throw new InvalidObjectException("Parent table not found: " + tableName);
-      }
-
-      MColumnDescriptor parentCD = retrieveCD ? nParentTable.mcd : parentTable.getSd().getCD();
-      int parentIntegerIndex = getColumnIndexFromTableColumns(parentCD == null ? null : parentCD.getCols(), columnName);
-      if (parentIntegerIndex == -1) {
-        if (parentTable.getPartitionKeys() != null) {
-          parentCD = null;
-          parentIntegerIndex = getColumnIndexFromTableColumns(parentTable.getPartitionKeys(), columnName);
-        }
-        if (parentIntegerIndex == -1) {
-          throw new InvalidObjectException("Parent column not found: " + columnName);
-        }
-      }
-      if (uk.getUk_name() == null) {
-        if (uk.getKey_seq() == 1) {
-          constraintName = generateConstraintName(parentTable, tableDB, tableName, columnName, "uk");
-        }
-      } else {
-        constraintName = normalizeIdentifier(uk.getUk_name());
-        if (constraintNameAlreadyExists(parentTable, constraintName)) {
-          String fqConstraintName = String.format("%s.%s.%s", parentTable.getDatabase().getName(),
-              parentTable.getTableName(), constraintName);
-          throw new InvalidObjectException("Constraint name already exists: " + fqConstraintName);
-        }
-      }
-
-
-      int enableValidateRely = (uk.isEnable_cstr() ? 4 : 0) +
-          (uk.isValidate_cstr() ? 2 : 0) + (uk.isRely_cstr() ? 1 : 0);
-      MConstraint muk = new MConstraint(
-          constraintName,
-          uk.getKey_seq(),
-          MConstraint.UNIQUE_CONSTRAINT,
-          null,
-          null,
-          enableValidateRely,
-          parentTable,
-          null,
-          parentCD,
-          null,
-          null,
-          parentIntegerIndex);
-      cstrs.add(muk);
-
-      // Add normalized identifier back to result
-      uk.setCatName(catName);
-      uk.setTable_db(tableDB);
-      uk.setTable_name(tableName);
-      uk.setColumn_name(columnName);
-      uk.setUk_name(constraintName);
-
-    }
-    pm.makePersistentAll(cstrs);
-    return uks;
-  }
-
-  @Override
-  public List<SQLNotNullConstraint> addNotNullConstraints(List<SQLNotNullConstraint> nns)
-          throws InvalidObjectException, MetaException {
-    return addNotNullConstraints(nns, true);
-  }
-
-  @Override
-  public List<SQLDefaultConstraint> addDefaultConstraints(List<SQLDefaultConstraint> nns)
-      throws InvalidObjectException, MetaException {
-    return addDefaultConstraints(nns, true);
-  }
-
-  @Override
-  public List<SQLCheckConstraint> addCheckConstraints(List<SQLCheckConstraint> nns)
-      throws InvalidObjectException, MetaException {
-    return addCheckConstraints(nns, true);
-  }
-
-  private List<SQLCheckConstraint> addCheckConstraints(List<SQLCheckConstraint> ccs, boolean retrieveCD)
-      throws InvalidObjectException, MetaException {
-    List<MConstraint> cstrs = new ArrayList<>();
-
-    for (SQLCheckConstraint cc: ccs) {
-      final String catName = normalizeIdentifier(cc.getCatName());
-      final String tableDB = normalizeIdentifier(cc.getTable_db());
-      final String tableName = normalizeIdentifier(cc.getTable_name());
-      final String columnName = cc.getColumn_name() == null? null
-          : normalizeIdentifier(cc.getColumn_name());
-      final String ccName = cc.getDc_name();
-      boolean isEnable = cc.isEnable_cstr();
-      boolean isValidate = cc.isValidate_cstr();
-      boolean isRely = cc.isRely_cstr();
-      String constraintValue = cc.getCheck_expression();
-      MConstraint muk = addConstraint(catName, tableDB, tableName, columnName, ccName, isEnable, isRely, isValidate,
-                    MConstraint.CHECK_CONSTRAINT, constraintValue, retrieveCD);
-      cstrs.add(muk);
-
-      // Add normalized identifier back to result
-      cc.setCatName(catName);
-      cc.setTable_db(tableDB);
-      cc.setTable_name(tableName);
-      cc.setColumn_name(columnName);
-      cc.setDc_name(muk.getConstraintName());
-    }
-    pm.makePersistentAll(cstrs);
-    return ccs;
-  }
-
-  private MConstraint addConstraint(String catName, String tableDB, String tableName, String columnName, String ccName,
-                               boolean isEnable, boolean isRely, boolean isValidate, int constraintType,
-                               String constraintValue, boolean retrieveCD)
-      throws InvalidObjectException, MetaException {
-    String constraintName = null;
-    // If retrieveCD is false, we do not need to do a deep retrieval of the Table Column Descriptor.
-    // For instance, this is the case when we are creating the table.
-    AttachedMTableInfo nParentTable = getMTable(catName, tableDB, tableName, retrieveCD);
-    MTable parentTable = nParentTable.mtbl;
-    if (parentTable == null) {
-      throw new InvalidObjectException("Parent table not found: " + tableName);
-    }
-
-    MColumnDescriptor parentCD = retrieveCD ? nParentTable.mcd : parentTable.getSd().getCD();
-    int parentIntegerIndex = getColumnIndexFromTableColumns(parentCD == null ? null : parentCD.getCols(), columnName);
-    if (parentIntegerIndex == -1) {
-      if (parentTable.getPartitionKeys() != null) {
-        parentCD = null;
-        parentIntegerIndex = getColumnIndexFromTableColumns(parentTable.getPartitionKeys(), columnName);
-      }
-    }
-    if (ccName == null) {
-      constraintName = generateConstraintName(parentTable, tableDB, tableName, columnName, "dc");
-    } else {
-      constraintName = normalizeIdentifier(ccName);
-      if (constraintNameAlreadyExists(parentTable, constraintName)) {
-        String fqConstraintName = String.format("%s.%s.%s", parentTable.getDatabase().getName(),
-            parentTable.getTableName(), constraintName);
-        throw new InvalidObjectException("Constraint name already exists: " + fqConstraintName);
-      }
-    }
-
-    int enableValidateRely = (isEnable ? 4 : 0) +
-        (isValidate ? 2 : 0) + (isRely ? 1 : 0);
-    MConstraint muk = new MConstraint(
-        constraintName,
-        1,
-        constraintType, // Not null constraint should reference a single column
-        null,
-        null,
-        enableValidateRely,
-        parentTable,
-        null,
-        parentCD,
-        null,
-        null,
-        parentIntegerIndex,
-        constraintValue);
-
-    return muk;
-  }
-
-  private List<SQLDefaultConstraint> addDefaultConstraints(List<SQLDefaultConstraint> dcs, boolean retrieveCD)
-      throws InvalidObjectException, MetaException {
-
-    List<MConstraint> cstrs = new ArrayList<>();
-    for (SQLDefaultConstraint dc : dcs) {
-      final String catName = normalizeIdentifier(dc.getCatName());
-      final String tableDB = normalizeIdentifier(dc.getTable_db());
-      final String tableName = normalizeIdentifier(dc.getTable_name());
-      final String columnName = normalizeIdentifier(dc.getColumn_name());
-      final String dcName = dc.getDc_name();
-      boolean isEnable = dc.isEnable_cstr();
-      boolean isValidate = dc.isValidate_cstr();
-      boolean isRely = dc.isRely_cstr();
-      String constraintValue = dc.getDefault_value();
-      MConstraint muk = addConstraint(catName, tableDB, tableName, columnName, dcName, isEnable, isRely, isValidate,
-      MConstraint.DEFAULT_CONSTRAINT, constraintValue, retrieveCD);
-      cstrs.add(muk);
-
-      // Add normalized identifier back to result
-      dc.setCatName(catName);
-      dc.setTable_db(tableDB);
-      dc.setTable_name(tableName);
-      dc.setColumn_name(columnName);
-      dc.setDc_name(muk.getConstraintName());
-    }
-    pm.makePersistentAll(cstrs);
-    return dcs;
-  }
-
-  private List<SQLNotNullConstraint> addNotNullConstraints(List<SQLNotNullConstraint> nns, boolean retrieveCD)
-          throws InvalidObjectException, MetaException {
-
-    List<MConstraint> cstrs = new ArrayList<>();
-    String constraintName;
-
-    for (SQLNotNullConstraint nn : nns) {
-      final String catName = normalizeIdentifier(nn.getCatName());
-      final String tableDB = normalizeIdentifier(nn.getTable_db());
-      final String tableName = normalizeIdentifier(nn.getTable_name());
-      final String columnName = normalizeIdentifier(nn.getColumn_name());
-
-      // If retrieveCD is false, we do not need to do a deep retrieval of the Table Column Descriptor.
-      // For instance, this is the case when we are creating the table.
-      AttachedMTableInfo nParentTable = getMTable(catName, tableDB, tableName, retrieveCD);
-      MTable parentTable = nParentTable.mtbl;
-      if (parentTable == null) {
-        throw new InvalidObjectException("Parent table not found: " + tableName);
-      }
-
-      MColumnDescriptor parentCD = retrieveCD ? nParentTable.mcd : parentTable.getSd().getCD();
-      int parentIntegerIndex = getColumnIndexFromTableColumns(parentCD == null ? null : parentCD.getCols(), columnName);
-      if (parentIntegerIndex == -1) {
-        if (parentTable.getPartitionKeys() != null) {
-          parentCD = null;
-          parentIntegerIndex = getColumnIndexFromTableColumns(parentTable.getPartitionKeys(), columnName);
-        }
-        if (parentIntegerIndex == -1) {
-          throw new InvalidObjectException("Parent column not found: " + columnName);
-        }
-      }
-      if (nn.getNn_name() == null) {
-        constraintName = generateConstraintName(parentTable, tableDB, tableName, columnName, "nn");
-      } else {
-        constraintName = normalizeIdentifier(nn.getNn_name());
-        if (constraintNameAlreadyExists(parentTable, constraintName)) {
-          String fqConstraintName = String.format("%s.%s.%s", parentTable.getDatabase().getName(),
-              parentTable.getTableName(), constraintName);
-          throw new InvalidObjectException("Constraint name already exists: " + fqConstraintName);
-        }
-      }
-
-      int enableValidateRely = (nn.isEnable_cstr() ? 4 : 0) +
-          (nn.isValidate_cstr() ? 2 : 0) + (nn.isRely_cstr() ? 1 : 0);
-      MConstraint muk = new MConstraint(
-          constraintName,
-          1,
-          MConstraint.NOT_NULL_CONSTRAINT, // Not null constraint should reference a single column
-          null,
-          null,
-          enableValidateRely,
-          parentTable,
-          null,
-          parentCD,
-          null,
-          null,
-          parentIntegerIndex);
-      cstrs.add(muk);
-      // Add normalized identifier back to result
-      nn.setCatName(catName);
-      nn.setTable_db(tableDB);
-      nn.setTable_name(tableName);
-      nn.setColumn_name(columnName);
-      nn.setNn_name(constraintName);
-    }
-    pm.makePersistentAll(cstrs);
-    return nns;
-  }
-
-  private void writeMTableColumnStatistics(Table table, MTableColumnStatistics mStatsObj,
-      MTableColumnStatistics oldStats) throws MetaException {
-
-    Preconditions.checkState(this.currentTransaction.isActive());
-
-    String colName = mStatsObj.getColName();
-
-    LOG.info("Updating table level column statistics for table={} colName={}",
-        Warehouse.getCatalogQualifiedTableName(table), colName);
-    validateTableCols(table, Lists.newArrayList(colName));
-
-    if (oldStats != null) {
-      StatObjectConverter.setFieldsIntoOldStats(mStatsObj, oldStats);
-    } else {
-      pm.makePersistent(mStatsObj);
-    }
-  }
-
-  private void writeMPartitionColumnStatistics(Table table, Partition partition,
-      MPartitionColumnStatistics mStatsObj, MPartitionColumnStatistics oldStats) {
-    String catName = mStatsObj.getPartition().getTable().getDatabase().getCatalogName();
-    String dbName = mStatsObj.getPartition().getTable().getDatabase().getName();
-    String tableName = mStatsObj.getPartition().getTable().getTableName();
-    String partName = mStatsObj.getPartition().getPartitionName();
-    String colName = mStatsObj.getColName();
-
-    Preconditions.checkState(this.currentTransaction.isActive());
-
-    LOG.info("Updating partition level column statistics for table=" +
-        TableName.getQualified(catName, dbName, tableName) +
-        " partName=" + partName + " colName=" + colName);
-
-    boolean foundCol = false;
-    List<FieldSchema> colList = partition.getSd().getCols();
-    for (FieldSchema col : colList) {
-      if (col.getName().equals(mStatsObj.getColName())) {
-        foundCol = true;
-        break;
-      }
-    }
-
-    if (!foundCol) {
-      LOG.warn("Column " + colName + " for which stats gathering is requested doesn't exist.");
-    }
-
-    if (oldStats != null) {
-      StatObjectConverter.setFieldsIntoOldStats(mStatsObj, oldStats);
-    } else {
-      pm.makePersistent(mStatsObj);
-    }
-  }
-
-  @Override
-  public Map<String, String> updateTableColumnStatistics(ColumnStatistics colStats, String validWriteIds, long writeId)
-      throws NoSuchObjectException, MetaException, InvalidObjectException, InvalidInputException {
-    boolean committed = false;
-    List<ColumnStatisticsObj> statsObjs = colStats.getStatsObj();
-    ColumnStatisticsDesc statsDesc = colStats.getStatsDesc();
-    long start = System.currentTimeMillis();
-    String catName = statsDesc.isSetCatName() ? statsDesc.getCatName() : getDefaultCatalog(conf);
-    try {
-      openTransaction();
-      // DataNucleus objects get detached all over the place for no (real) reason.
-      // So let's not use them anywhere unless absolutely necessary.
-      MTable mTable = ensureGetMTable(catName, statsDesc.getDbName(), statsDesc.getTableName());
-      int maxRetries = MetastoreConf.getIntVar(conf, ConfVars.METASTORE_S4U_NOWAIT_MAX_RETRIES);
-      long sleepInterval = MetastoreConf.getTimeVar(conf,
-          ConfVars.METASTORE_S4U_NOWAIT_RETRY_SLEEP_INTERVAL, TimeUnit.MILLISECONDS);
-      Map<String, String> result = new RetryingExecutor<>(maxRetries, () -> {
-        AtomicReference<Exception> exceptionRef = new AtomicReference<>();
-        String savePoint = "uts_" + ThreadLocalRandom.current().nextInt(10000) + "_" + System.nanoTime();
-        setTransactionSavePoint(savePoint);
-        executePlainSQL(
-            sqlGenerator.addForUpdateNoWait("SELECT \"TBL_ID\" FROM \"TBLS\" WHERE \"TBL_ID\" = " + mTable.getId()),
-            true,
-            exception -> {
-              rollbackTransactionToSavePoint(savePoint);
-              exceptionRef.set(exception);
-            });
-        if (exceptionRef.get() != null) {
-          throw new RetryingExecutor.RetryException(exceptionRef.get());
-        }
-        pm.refresh(mTable);
-        Table table = convertToTable(mTable, conf);
-        List<String> colNames = new ArrayList<>();
-        for (ColumnStatisticsObj statsObj : statsObjs) {
-          colNames.add(statsObj.getColName());
-        }
-
-        Map<String, MTableColumnStatistics> oldStats = Maps.newHashMap();
-        List<MTableColumnStatistics> stats = getMTableColumnStatistics(table, colNames, colStats.getEngine());
-        for (MTableColumnStatistics cStat : stats) {
-          oldStats.put(cStat.getColName(), cStat);
-        }
-
-        for (ColumnStatisticsObj statsObj : statsObjs) {
-          MTableColumnStatistics mStatsObj = StatObjectConverter.convertToMTableColumnStatistics(mTable, statsDesc,
-              statsObj, colStats.getEngine());
-          writeMTableColumnStatistics(table, mStatsObj, oldStats.get(statsObj.getColName()));
-          // There is no need to add colname again, otherwise we will get duplicate colNames.
-        }
-
-        // Set the table properties
-        // No need to check again if it exists.
-        String dbname = table.getDbName();
-        String name = table.getTableName();
-        MTable oldt = mTable;
-        Map<String, String> newParams = new HashMap<>(table.getParameters());
-        StatsSetupConst.setColumnStatsState(newParams, colNames);
-        boolean isTxn = TxnUtils.isTransactionalTable(oldt.getParameters());
-        if (isTxn) {
-          if (!areTxnStatsSupported) {
-            StatsSetupConst.setBasicStatsState(newParams, StatsSetupConst.FALSE);
-          } else {
-            String errorMsg = verifyStatsChangeCtx(TableName.getDbTable(dbname, name), oldt.getParameters(), newParams,
-                writeId, validWriteIds, true);
-            if (errorMsg != null) {
-              throw new MetaException(errorMsg);
-            }
-            if (!isCurrentStatsValidForTheQuery(oldt, validWriteIds, true)) {
-              // Make sure we set the flag to invalid regardless of the current value.
-              StatsSetupConst.setBasicStatsState(newParams, StatsSetupConst.FALSE);
-              LOG.info("Removed COLUMN_STATS_ACCURATE from the parameters of the table " + dbname + "." + name);
-            }
-            oldt.setWriteId(writeId);
-          }
-        }
-        oldt.setParameters(newParams);
-        return newParams;
-      }).onRetry(e -> e instanceof RetryingExecutor.RetryException)
-        .commandName("updateTableColumnStatistics").sleepInterval(sleepInterval, interval ->
-              ThreadLocalRandom.current().nextLong(sleepInterval) + 30).run();
-      committed = commitTransaction();
-      return committed ? result : null;
-    } finally {
-      LOG.debug("{} updateTableColumnStatistics took {}ms, success: {}",
-          new TableName(catName, statsDesc.getDbName(), statsDesc.getTableName()),
-          System.currentTimeMillis() - start, committed);
-      rollbackAndCleanup(committed, null);
-    }
-  }
-
-  @Override
-  public Map<String, String> updatePartitionColumnStatistics(Table table, MTable mTable, ColumnStatistics colStats,
-      List<String> partVals, String validWriteIds, long writeId)
-          throws MetaException, NoSuchObjectException, InvalidObjectException, InvalidInputException {
-    boolean committed = false;
-    long start = System.currentTimeMillis();
-    List<ColumnStatisticsObj> statsObjs = colStats.getStatsObj();
-    ColumnStatisticsDesc statsDesc = colStats.getStatsDesc();
-    String catName = statsDesc.isSetCatName() ? statsDesc.getCatName() : getDefaultCatalog(conf);
-    try {
-      openTransaction();
-      MPartition mPartition =
-          ensureGetMPartition(new TableName(catName, statsDesc.getDbName(), statsDesc.getTableName()), partVals);
-      if (mPartition == null) {
-        throw new NoSuchObjectException("Partition for which stats is gathered doesn't exist.");
-      }
-
-      List<String> colNames = new ArrayList<>();
-      for(ColumnStatisticsObj statsObj : statsObjs) {
-        colNames.add(statsObj.getColName());
-      }
-      int maxRetries = MetastoreConf.getIntVar(conf, ConfVars.METASTORE_S4U_NOWAIT_MAX_RETRIES);
-      long sleepInterval = MetastoreConf.getTimeVar(conf,
-          ConfVars.METASTORE_S4U_NOWAIT_RETRY_SLEEP_INTERVAL, TimeUnit.MILLISECONDS);
-      Map<String, String> result = new RetryingExecutor<>(maxRetries, () -> {
-        AtomicReference<Exception> exceptionRef = new AtomicReference<>();
-        String savePoint = "ups_" + ThreadLocalRandom.current().nextInt(10000) + "_" + System.nanoTime();
-        setTransactionSavePoint(savePoint);
-        executePlainSQL(sqlGenerator.addForUpdateNoWait(
-            "SELECT \"PART_ID\" FROM \"PARTITIONS\" WHERE \"PART_ID\" = " + mPartition.getId()),
-            true,
-            exception -> {
-              rollbackTransactionToSavePoint(savePoint);
-              exceptionRef.set(exception);
-            });
-        if (exceptionRef.get() != null) {
-          throw new RetryingExecutor.RetryException(exceptionRef.get());
-        }
-        pm.refresh(mPartition);
-        Partition partition = convertToPart(catName, statsDesc.getDbName(), statsDesc.getTableName(),
-            mPartition, TxnUtils.isAcidTable(table), conf);
-        Map<String, MPartitionColumnStatistics> oldStats = Maps.newHashMap();
-        List<MPartitionColumnStatistics> stats =
-            getMPartitionColumnStatistics(table, Lists.newArrayList(statsDesc.getPartName()), colNames, colStats.getEngine());
-        for (MPartitionColumnStatistics cStat : stats) {
-          oldStats.put(cStat.getColName(), cStat);
-        }
-
-        for (ColumnStatisticsObj statsObj : statsObjs) {
-          MPartitionColumnStatistics mStatsObj = StatObjectConverter.convertToMPartitionColumnStatistics(mPartition,
-              statsDesc, statsObj, colStats.getEngine());
-          writeMPartitionColumnStatistics(table, partition, mStatsObj, oldStats.get(statsObj.getColName()));
-        }
-
-        Map<String, String> newParams = new HashMap<>(mPartition.getParameters());
-        StatsSetupConst.setColumnStatsState(newParams, colNames);
-        boolean isTxn = TxnUtils.isTransactionalTable(table);
-        if (isTxn) {
-          if (!areTxnStatsSupported) {
-            StatsSetupConst.setBasicStatsState(newParams, StatsSetupConst.FALSE);
-          } else {
-            String errorMsg = verifyStatsChangeCtx(
-                TableName.getDbTable(statsDesc.getDbName(), statsDesc.getTableName()), mPartition.getParameters(),
-                newParams, writeId, validWriteIds, true);
-            if (errorMsg != null) {
-              throw new MetaException(errorMsg);
-            }
-            if (!isCurrentStatsValidForTheQuery(mPartition.getParameters(), mPartition.getWriteId(), validWriteIds, true)) {
-              // Make sure we set the flag to invalid regardless of the current value.
-              StatsSetupConst.setBasicStatsState(newParams, StatsSetupConst.FALSE);
-              LOG.info("Removed COLUMN_STATS_ACCURATE from the parameters of the partition: {}, {} ",
-                  new TableName(catName, statsDesc.getDbName(), statsDesc.getTableName()), statsDesc.getPartName());
-            }
-            mPartition.setWriteId(writeId);
-          }
-        }
-        mPartition.setParameters(newParams);
-        return newParams;
-      }).onRetry(e -> e instanceof RetryingExecutor.RetryException)
-          .commandName("updatePartitionColumnStatistics").sleepInterval(sleepInterval, interval ->
-              ThreadLocalRandom.current().nextLong(sleepInterval) + 30).run();
-      committed = commitTransaction();
-      return committed ? result : null;
-    } finally {
-      LOG.debug("{} updatePartitionColumnStatistics took {}ms, success: {}",
-          new TableName(catName, statsDesc.getDbName(), statsDesc.getTableName()),
-          System.currentTimeMillis() - start, committed);
-      rollbackAndCleanup(committed, null);
-    }
-  }
-
-  @Override
-  public Map<String, String> updatePartitionColumnStatistics(ColumnStatistics colStats,
-      List<String> partVals, String validWriteIds, long writeId)
-      throws MetaException, NoSuchObjectException, InvalidObjectException, InvalidInputException {
-    ColumnStatisticsDesc statsDesc = colStats.getStatsDesc();
-    Table table = getTable(statsDesc.getCatName(), statsDesc.getDbName(), statsDesc.getTableName());
-    MTable mTable = ensureGetMTable(statsDesc.getCatName(), statsDesc.getDbName(), statsDesc.getTableName());
-    return updatePartitionColumnStatistics(table, mTable, colStats, partVals, validWriteIds, writeId);
-  }
-
-  @Override
-  public Map<String, Map<String, String>> updatePartitionColumnStatisticsInBatch(
-                                                      Map<String, ColumnStatistics> partColStatsMap,
-                                                      Table tbl,
-                                                      List<TransactionalMetaStoreEventListener> listeners,
-                                                      String validWriteIds, long writeId)
-          throws NoSuchObjectException, MetaException, InvalidObjectException, InvalidInputException {
-
-    return new GetHelper<Map<String, Map<String, String>>>(tbl.getCatName(),
-        tbl.getDbName(), tbl.getTableName(), true, false) {
-      @Override
-      protected String describeResult() {
-        return "Map of partition key to column stats if successful";
-      }
-      @Override
-      protected Map<String, Map<String, String>> getSqlResult(GetHelper<Map<String, Map<String, String>>> ctx)
-          throws MetaException {
-        return directSql.updatePartitionColumnStatisticsBatch(partColStatsMap, tbl,
-            listeners, validWriteIds, writeId);
-      }
-      @Override
-      protected Map<String, Map<String, String>> getJdoResult(GetHelper<Map<String, Map<String, String>>> ctx)
-          throws MetaException, NoSuchObjectException, InvalidObjectException, InvalidInputException {
-        throw new UnsupportedOperationException("Cannot update partition column statistics with JDO, make sure direct SQL is enabled");
-      }
-    }.run(false);
-  }
-
-  private List<MTableColumnStatistics> getMTableColumnStatistics(Table table, List<String> colNames, String engine)
-      throws MetaException {
-
-    Preconditions.checkState(this.currentTransaction.isActive());
-
-    if (colNames.isEmpty()) {
-      return Collections.emptyList();
-    }
-
-    boolean committed = false;
-    try {
-      openTransaction();
-
-      validateTableCols(table, colNames);
-
-      List<MTableColumnStatistics> result = Collections.emptyList();
-      try (Query query = pm.newQuery(MTableColumnStatistics.class)) {
-      result =
-          Batchable.runBatched(batchSize, colNames, new Batchable<String, MTableColumnStatistics>() {
-            @Override
-            public List<MTableColumnStatistics> run(List<String> input)
-                throws MetaException {
-              StringBuilder filter =
-                  new StringBuilder("table.tableName == t1 && table.database.name == t2 && table.database.catalogName == t3 && engine == t4 && (");
-              StringBuilder paramStr = new StringBuilder(
-                  "java.lang.String t1, java.lang.String t2, java.lang.String t3, java.lang.String t4");
-              Object[] params = new Object[input.size() + 4];
-              params[0] = table.getTableName();
-              params[1] = table.getDbName();
-              params[2] = table.getCatName();
-              params[3] = engine;
-              for (int i = 0; i < input.size(); ++i) {
-                filter.append((i == 0) ? "" : " || ").append("colName == c").append(i);
-                paramStr.append(", java.lang.String c").append(i);
-                params[i + 4] = input.get(i);
-              }
-              filter.append(")");
-              query.setFilter(filter.toString());
-              query.declareParameters(paramStr.toString());
-              List<MTableColumnStatistics> paritial = (List<MTableColumnStatistics>) query.executeWithArray(params);
-              pm.retrieveAll(paritial);
-              return paritial;
-            }
-          });
-
-      if (result.size() > colNames.size()) {
-        throw new MetaException("Unexpected " + result.size() + " statistics for "
-            + colNames.size() + " columns");
-      }
-      result = new ArrayList<>(result);
-      }
-      committed = commitTransaction();
-      return result;
-    } catch (Exception ex) {
-      LOG.error("Error retrieving statistics via jdo", ex);
-      if (ex instanceof MetaException) {
-        throw (MetaException) ex;
-      }
-      throw new MetaException(ex.getMessage());
-    } finally {
-      rollbackAndCleanup(committed, null);
-    }
-  }
-
-  @VisibleForTesting
-  public void validateTableCols(Table table, List<String> colNames) throws MetaException {
-    List<FieldSchema> colList = table.getSd().getCols();
-    for (String colName : colNames) {
-      boolean foundCol = false;
-      for (FieldSchema mCol : colList) {
-        if (mCol.getName().equals(colName)) {
-          foundCol = true;
-          break;
-        }
-      }
-      if (!foundCol) {
-        throw new MetaException("Column " + colName + " doesn't exist in table "
-            + table.getTableName() + " in database " + table.getDbName());
-      }
-    }
-  }
-
-  @Override
-  public List<ColumnStatistics> getTableColumnStatistics(
-      String catName,
-      String dbName,
-      String tableName,
-      List<String> colNames) throws MetaException, NoSuchObjectException {
-    // Note: this will get stats without verifying ACID.
-    boolean committed = false;
-    Query query = null;
-    List<ColumnStatistics> result = new ArrayList<>();
-
-    try {
-      openTransaction();
-      query = pm.newQuery(MTableColumnStatistics.class);
-      query.setFilter("table.tableName == t1 && table.database.name == t2 && table.database.catalogName == t3");
-      query.declareParameters("java.lang.String t1, java.lang.String t2, java.lang.String t3");
-      query.setResult("DISTINCT engine");
-      Collection names = (Collection) query.execute(tableName, dbName, catName);
-      List<String> engines = new ArrayList<>();
-      for (Iterator i = names.iterator(); i.hasNext();) {
-        engines.add((String) i.next());
-      }
-      for (String e : engines) {
-        ColumnStatistics cs = getTableColumnStatisticsInternal(
-            catName, dbName, tableName, colNames, e, true, true);
-        if (cs != null) {
-          result.add(cs);
-        }
-      }
-      committed = commitTransaction();
-      return result;
-    } finally {
-      LOG.debug("Done executing getTableColumnStatistics with status : {}",
-          committed);
-      rollbackAndCleanup(committed, query);
-    }
-  }
-
-  @Override
-  public ColumnStatistics getTableColumnStatistics(
-      String catName,
-      String dbName,
-      String tableName,
-      List<String> colNames,
-      String engine) throws MetaException, NoSuchObjectException {
-    // Note: this will get stats without verifying ACID.
-    return getTableColumnStatisticsInternal(
-        catName, dbName, tableName, colNames, engine, true, true);
-  }
-
-  @Override
-  public ColumnStatistics getTableColumnStatistics(
-      String catName,
-      String dbName,
-      String tableName,
-      List<String> colNames,
-      String engine,
-      String writeIdList) throws MetaException, NoSuchObjectException {
-    // If the current stats in the metastore doesn't comply with
-    // the isolation level of the query, set No to the compliance flag.
-    Boolean isCompliant = null;
-    if (writeIdList != null) {
-      MTable table = this.getMTable(catName, dbName, tableName);
-      if (table == null) {
-        throw new NoSuchObjectException(TableName.getQualified(catName, dbName, tableName) + " table not found");
-      }
-      isCompliant = !TxnUtils.isTransactionalTable(table.getParameters())
-        || (areTxnStatsSupported && isCurrentStatsValidForTheQuery(table, writeIdList, false));
-    }
-    ColumnStatistics stats = getTableColumnStatisticsInternal(
-        catName, dbName, tableName, colNames, engine, true, true);
-    if (stats != null && isCompliant != null) {
-      stats.setIsStatsCompliant(isCompliant);
-    }
-    return stats;
-  }
-
-  protected ColumnStatistics getTableColumnStatisticsInternal(
-      String catName, String dbName, String tableName, final List<String> colNames, String engine,
-      boolean allowSql, boolean allowJdo) throws MetaException, NoSuchObjectException {
-    final boolean enableBitVector = MetastoreConf.getBoolVar(getConf(), ConfVars.STATS_FETCH_BITVECTOR);
-    final boolean enableKll = MetastoreConf.getBoolVar(getConf(), ConfVars.STATS_FETCH_KLL);
-    return new GetStatHelper(normalizeIdentifier(catName), normalizeIdentifier(dbName),
-        normalizeIdentifier(tableName), allowSql, allowJdo, null) {
-      @Override
-      protected ColumnStatistics getSqlResult(GetHelper<ColumnStatistics> ctx) throws MetaException {
-        return directSqlAggrStats.getTableStats(catName, dbName, tblName, colNames, engine, enableBitVector, enableKll);
-      }
-
-      @Override
-      protected ColumnStatistics getJdoResult(GetHelper<ColumnStatistics> ctx) throws MetaException {
-
-        List<MTableColumnStatistics> mStats = getMTableColumnStatistics(getTable(), colNames, engine);
-        if (mStats.isEmpty()) {
-          return null;
-        }
-        // LastAnalyzed is stored per column, but thrift object has it per
-        // multiple columns. Luckily, nobody actually uses it, so we will set to
-        // lowest value of all columns for now.
-        ColumnStatisticsDesc desc = StatObjectConverter.getTableColumnStatisticsDesc(mStats.get(0));
-        List<ColumnStatisticsObj> statObjs = new ArrayList<>(mStats.size());
-        for (MTableColumnStatistics mStat : mStats) {
-          if (desc.getLastAnalyzed() > mStat.getLastAnalyzed()) {
-            desc.setLastAnalyzed(mStat.getLastAnalyzed());
-          }
-          statObjs.add(StatObjectConverter.getColumnStatisticsObj(mStat, enableBitVector, enableKll));
-          Deadline.checkTimeout();
-        }
-        ColumnStatistics colStat = new ColumnStatistics(desc, statObjs);
-        colStat.setEngine(engine);
-        return colStat;
-      }
-    }.run(true);
-  }
-
-  @Override
-  public List<List<ColumnStatistics>> getPartitionColumnStatistics(String catName, String dbName, String tableName,
-      List<String> partNames, List<String> colNames) throws MetaException, NoSuchObjectException {
-    // Note: this will get stats without verifying ACID.
-    boolean committed = false;
-    Query query = null;
-    List<List<ColumnStatistics>> result = new ArrayList<>();
-
-    try {
-      openTransaction();
-      query = pm.newQuery(MPartitionColumnStatistics.class);
-      query.setFilter("partition.table.tableName == t1 && partition.table.database.name == t2 && partition.table.database.catalogName == t3");
-      query.declareParameters("java.lang.String t1, java.lang.String t2, java.lang.String t3");
-      query.setResult("DISTINCT engine");
-      Collection names = (Collection) query.execute(tableName, dbName, catName);
-      List<String> engines = new ArrayList<>();
-      for (Iterator i = names.iterator(); i.hasNext();) {
-        engines.add((String) i.next());
-      }
-      for (String e : engines) {
-        List<ColumnStatistics> cs = getPartitionColumnStatisticsInternal(
-            catName, dbName, tableName, partNames, colNames, e, true, true);
-        if (cs != null) {
-          result.add(cs);
-        }
-      }
-      committed = commitTransaction();
-      return result;
-    } finally {
-      LOG.debug("Done executing getTableColumnStatistics with status : {}",
-          committed);
-      rollbackAndCleanup(committed, query);
-    }
-  }
-
-  @Override
-  public List<ColumnStatistics> getPartitionColumnStatistics(String catName, String dbName, String tableName,
-      List<String> partNames, List<String> colNames, String engine) throws MetaException, NoSuchObjectException {
-    // Note: this will get stats without verifying ACID.
-    if (CollectionUtils.isEmpty(partNames) || CollectionUtils.isEmpty(colNames)) {
-      LOG.debug("PartNames and/or ColNames are empty");
-      return Collections.emptyList();
-    }
-    return getPartitionColumnStatisticsInternal(
-        catName, dbName, tableName, partNames, colNames, engine, true, true);
-  }
-
-  @Override
-  public List<ColumnStatistics> getPartitionColumnStatistics(
-      String catName, String dbName, String tableName,
-      List<String> partNames, List<String> colNames,
-      String engine, String writeIdList)
-      throws MetaException, NoSuchObjectException {
-    if (CollectionUtils.isEmpty(partNames) || CollectionUtils.isEmpty(colNames)) {
-      LOG.debug("PartNames and/or ColNames are empty");
-      return Collections.emptyList();
-    }
-    List<ColumnStatistics> allStats = getPartitionColumnStatisticsInternal(
-        catName, dbName, tableName, partNames, colNames, engine, true, true);
-    if (writeIdList != null) {
-      if (!areTxnStatsSupported) {
-        for (ColumnStatistics cs : allStats) {
-          cs.setIsStatsCompliant(false);
-        }
-      } else {
-        // TODO: this could be improved to get partitions in bulk
-        for (ColumnStatistics cs : allStats) {
-          MPartition mpart = ensureGetMPartition(new TableName(catName, dbName, tableName),
-              Warehouse.getPartValuesFromPartName(cs.getStatsDesc().getPartName()));
-          if (mpart == null
-              || !isCurrentStatsValidForTheQuery(mpart.getParameters(), mpart.getWriteId(), writeIdList, false)) {
-            if (mpart != null) {
-              LOG.debug("The current metastore transactional partition column statistics for {}.{}.{} "
-                + "(write ID {}) are not valid for current query ({} {})", dbName, tableName,
-                mpart.getPartitionName(), mpart.getWriteId(), writeIdList);
-            }
-            cs.setIsStatsCompliant(false);
-          } else {
-            cs.setIsStatsCompliant(true);
-          }
-        }
-      }
-    }
-    return allStats;
-  }
-
-  protected List<ColumnStatistics> getPartitionColumnStatisticsInternal(
-      String catName, String dbName, String tableName, final List<String> partNames, final List<String> colNames,
-      String engine, boolean allowSql, boolean allowJdo) throws MetaException, NoSuchObjectException {
-    final boolean enableBitVector = MetastoreConf.getBoolVar(getConf(), ConfVars.STATS_FETCH_BITVECTOR);
-    final boolean enableKll = MetastoreConf.getBoolVar(getConf(), ConfVars.STATS_FETCH_KLL);
-    return new GetListHelper<ColumnStatistics>(catName, dbName, tableName, allowSql, allowJdo) {
-      @Override
-      protected List<ColumnStatistics> getSqlResult(
-          GetHelper<List<ColumnStatistics>> ctx) throws MetaException {
-        return directSqlAggrStats.getPartitionStats(
-            catName, dbName, tblName, partNames, colNames, engine, enableBitVector, enableKll);
-      }
-      @Override
-      protected List<ColumnStatistics> getJdoResult(GetHelper<List<ColumnStatistics>> ctx)
-          throws MetaException, NoSuchObjectException {
-        List<MPartitionColumnStatistics> mStats =
-            getMPartitionColumnStatistics(getTable(), partNames, colNames, engine);
-        List<ColumnStatistics> result = new ArrayList<>(Math.min(mStats.size(), partNames.size()));
-        String lastPartName = null;
-        List<ColumnStatisticsObj> curList = null;
-        ColumnStatisticsDesc csd = null;
-        for (int i = 0; i <= mStats.size(); ++i) {
-          boolean isLast = i == mStats.size();
-          MPartitionColumnStatistics mStatsObj = isLast ? null : mStats.get(i);
-          String partName = isLast ? null : mStatsObj.getPartition().getPartitionName();
-          if (isLast || !partName.equals(lastPartName)) {
-            if (i != 0) {
-              ColumnStatistics colStat = new ColumnStatistics(csd, curList);
-              colStat.setEngine(engine);
-              result.add(colStat);
-            }
-            if (isLast) {
-              continue;
-            }
-            csd = StatObjectConverter.getPartitionColumnStatisticsDesc(mStatsObj);
-            curList = new ArrayList<>(colNames.size());
-          }
-          curList.add(StatObjectConverter.getColumnStatisticsObj(mStatsObj, enableBitVector, enableKll));
-          lastPartName = partName;
-          Deadline.checkTimeout();
-        }
-        return result;
-      }
-    }.run(true);
-  }
-
-  @Override
-  public AggrStats get_aggr_stats_for(String catName, String dbName, String tblName,
-      final List<String> partNames, final List<String> colNames,
-      String engine, String writeIdList) throws MetaException, NoSuchObjectException {
-    // If the current stats in the metastore doesn't comply with
-    // the isolation level of the query, return null.
-    if (writeIdList != null) {
-      if (partNames == null || partNames.isEmpty()) {
-        return null;
-      }
-
-      Table table = getTable(catName, dbName, tblName);
-      boolean isTxn = TxnUtils.isTransactionalTable(table.getParameters());
-      if (isTxn && !areTxnStatsSupported) {
-        return null;
-      }
-      GetPartitionsFilterSpec fs = new GetPartitionsFilterSpec();
-      fs.setFilterMode(PartitionFilterMode.BY_NAMES);
-      fs.setFilters(partNames);
-      GetProjectionsSpec ps = new GetProjectionsSpec();
-      ps.setIncludeParamKeyPattern(StatsSetupConst.COLUMN_STATS_ACCURATE + '%');
-      ps.setFieldList(Lists.newArrayList("writeId", "parameters", "values"));
-      List<Partition> parts = getPartitionSpecsByFilterAndProjection(table, ps, fs);
-
-      // Loop through the given "partNames" list
-      // checking isolation-level-compliance of each partition column stats.
-      for (Partition part : parts) {
-
-        if (!isCurrentStatsValidForTheQuery(part.getParameters(), part.getWriteId(), writeIdList, false)) {
-          String partName = Warehouse.makePartName(table.getPartitionKeys(), part.getValues());
-          LOG.debug("The current metastore transactional partition column "
-              + "statistics for {}.{}.{} is not valid for the current query",
-              dbName, tblName, partName);
-          return null;
-        }
-      }
-    }
-    return get_aggr_stats_for(catName, dbName, tblName, partNames, colNames, engine);
-  }
-
-  @Override
-  public AggrStats get_aggr_stats_for(String catName, String dbName, String tblName,
-      final List<String> partNames, final List<String> colNames, String engine)
-      throws MetaException, NoSuchObjectException {
-    final boolean useDensityFunctionForNDVEstimation = MetastoreConf.getBoolVar(getConf(),
-        ConfVars.STATS_NDV_DENSITY_FUNCTION);
-    final double ndvTuner = MetastoreConf.getDoubleVar(getConf(), ConfVars.STATS_NDV_TUNER);
-    final boolean enableBitVector = MetastoreConf.getBoolVar(getConf(), ConfVars.STATS_FETCH_BITVECTOR);
-    final boolean enableKll = MetastoreConf.getBoolVar(getConf(), ConfVars.STATS_FETCH_KLL);
-    return new GetHelper<AggrStats>(catName, dbName, tblName, true, false) {
-      @Override
-      protected AggrStats getSqlResult(GetHelper<AggrStats> ctx)
-          throws MetaException {
-        return directSql.aggrColStatsForPartitions(catName, dbName, tblName, partNames,
-            colNames, engine, useDensityFunctionForNDVEstimation, ndvTuner, enableBitVector, enableKll);
-      }
-      @Override
-      protected AggrStats getJdoResult(GetHelper<AggrStats> ctx)
-          throws MetaException, NoSuchObjectException {
-        // This is fast path for query optimizations, if we can find this info
-        // quickly using
-        // directSql, do it. No point in failing back to slow path here.
-        throw new MetaException("Jdo path is not implemented for stats aggr.");
-      }
-      @Override
-      protected String describeResult() {
-        return null;
-      }
-    }.run(true);
-  }
-
-  @Override
-  public List<MetaStoreServerUtils.ColStatsObjWithSourceInfo> getPartitionColStatsForDatabase(String catName, String dbName)
-      throws MetaException, NoSuchObjectException {
-    final boolean enableBitVector = MetastoreConf.getBoolVar(getConf(), ConfVars.STATS_FETCH_BITVECTOR);
-    final boolean enableKll = MetastoreConf.getBoolVar(getConf(), ConfVars.STATS_FETCH_KLL);
-    return new GetHelper<List<MetaStoreServerUtils.ColStatsObjWithSourceInfo>>(
-        catName, dbName, null, true, false) {
-      @Override
-      protected List<MetaStoreServerUtils.ColStatsObjWithSourceInfo> getSqlResult(
-          GetHelper<List<MetaStoreServerUtils.ColStatsObjWithSourceInfo>> ctx) throws MetaException {
-        return directSqlAggrStats.getColStatsForAllTablePartitions(catName, dbName, enableBitVector, enableKll);
-      }
-
-      @Override
-      protected List<MetaStoreServerUtils.ColStatsObjWithSourceInfo> getJdoResult(
-          GetHelper<List<MetaStoreServerUtils.ColStatsObjWithSourceInfo>> ctx)
-          throws MetaException, NoSuchObjectException {
-        // This is fast path for query optimizations, if we can find this info
-        // quickly using directSql, do it. No point in failing back to slow path
-        // here.
-        throw new MetaException("Jdo path is not implemented for getPartitionColStatsForDatabase.");
-      }
-
-      @Override
-      protected String describeResult() {
-        return null;
-      }
-    }.run(true);
-  }
-
-  private List<MPartitionColumnStatistics> getMPartitionColumnStatistics(Table table, List<String> partNames,
-      List<String> colNames, String engine) throws MetaException {
-    boolean committed = false;
-
-    try {
-      openTransaction();
-      // We are not going to verify SD for each partition. Just verify for the
-      // table. TODO: we need verify the partition column instead
-      try {
-        validateTableCols(table, colNames);
-      } catch (MetaException me) {
-        LOG.warn("The table does not have the same column definition as its partition.");
-      }
-      List<MPartitionColumnStatistics> result = Collections.emptyList();
-      try (Query query = pm.newQuery(MPartitionColumnStatistics.class)) {
-        String paramStr = "java.lang.String t1, java.lang.String t2, java.lang.String t3, java.lang.String t4";
-        String filter = "partition.table.tableName == t1 && partition.table.database.name == t2 && partition.table.database.catalogName == t3 && engine == t4 && (";
-        Object[] params = new Object[colNames.size() + partNames.size() + 4];
-        int i = 0;
-        params[i++] = table.getTableName();
-        params[i++] = table.getDbName();
-        params[i++] = table.isSetCatName() ? table.getCatName() : getDefaultCatalog(conf);
-        params[i++] = engine;
-        int firstI = i;
-        for (String s : partNames) {
-          filter += ((i == firstI) ? "" : " || ") + "partition.partitionName == p" + i;
-          paramStr += ", java.lang.String p" + i;
-          params[i++] = s;
-        }
-        filter += ") && (";
-        firstI = i;
-        for (String s : colNames) {
-          filter += ((i == firstI) ? "" : " || ") + "colName == c" + i;
-          paramStr += ", java.lang.String c" + i;
-          params[i++] = s;
-        }
-        filter += ")";
-        query.setFilter(filter);
-        query.declareParameters(paramStr);
-        query.setOrdering("partition.partitionName ascending");
-        result = (List<MPartitionColumnStatistics>) query.executeWithArray(params);
-        pm.retrieveAll(result);
-        result = new ArrayList<>(result);
-      } catch (Exception ex) {
-        LOG.error("Error retrieving statistics via jdo", ex);
-        throw new MetaException(ex.getMessage());
-      }
-      committed = commitTransaction();
-      return result;
-    } finally {
-      if (!committed) {
-        rollbackTransaction();
-        return Collections.emptyList();
-      }
-    }
-  }
-
-  @Override
-  public void deleteAllPartitionColumnStatistics(TableName tn, String writeIdList) {
-
-    String catName = tn.getCat();
-    String dbName = tn.getDb();
-    String tableName = tn.getTable();
-
-    Query query = null;
-    dbName = org.apache.commons.lang3.StringUtils.defaultString(dbName, Warehouse.DEFAULT_DATABASE_NAME);
-    catName = normalizeIdentifier(catName);
-    if (tableName == null) {
-      throw new RuntimeException("Table name is null.");
-    }
-    boolean ret = false;
-    try {
-      openTransaction();
-      MTable mTable = getMTable(catName, dbName, tableName);
-
-      query = pm.newQuery(MPartitionColumnStatistics.class);
-
-      String filter = "partition.table.database.name == t2 && partition.table.tableName == t3 && partition.table.database.catalogName == t4";
-      String parameters = "java.lang.String t2, java.lang.String t3, java.lang.String t4";
-
-      query.setFilter(filter);
-      query.declareParameters(parameters);
-
-      Long number = query.deletePersistentAll(normalizeIdentifier(dbName), normalizeIdentifier(tableName),
-          normalizeIdentifier(catName));
-
-      new GetHelper<Integer>(catName, dbName, tableName, true, true) {
-        private final SqlFilterForPushdown filter = new SqlFilterForPushdown();
-
-        @Override
-        protected String describeResult() {
-          return "Partition count";
-        }
-
-        @Override
-        protected boolean canUseDirectSql(GetHelper<Integer> ctx) throws MetaException {
-          return true;
-        }
-
-        @Override
-        protected Integer getSqlResult(GetHelper<Integer> ctx) throws MetaException {
-          directSql.deleteColumnStatsState(getTable().getId());
-          return 0;
-        }
-
-        @Override
-        protected Integer getJdoResult(GetHelper<Integer> ctx) throws MetaException, NoSuchObjectException {
-          try {
-            List<Partition> parts = getPartitions(catName, dbName, tableName,
-                GetPartitionsArgs.getAllPartitions());
-            for (Partition part : parts) {
-              Partition newPart = new Partition(part);
-              StatsSetupConst.clearColumnStatsState(newPart.getParameters());
-              alterPartition(catName, dbName, tableName, part.getValues(), newPart, writeIdList);
-            }
-            return parts.size();
-          } catch (InvalidObjectException e) {
-            LOG.error("error updating parts", e);
-            return -1;
-          }
-        }
-      }.run(true);
-
-      ret = commitTransaction();
-    } catch (Exception e) {
-      LOG.error("Couldn't clear stats for table", e);
-    } finally {
-      rollbackAndCleanup(ret, query);
-    }
-  }
-
-  @Override
-  public boolean deletePartitionColumnStatistics(String catName, String dbName, String tableName,
-      List<String> partNames, List<String> colNames, String engine)
-      throws NoSuchObjectException, MetaException, InvalidObjectException, InvalidInputException {
-    if (partNames == null || partNames.isEmpty()) {
-      throw new InvalidInputException("No partition specified for dropping the statistics");
-    }
-    dbName = org.apache.commons.lang3.StringUtils.defaultString(dbName, Warehouse.DEFAULT_DATABASE_NAME);
-    catName = normalizeIdentifier(catName);
-    List<String> cols = normalizeIdentifiers(colNames);
-    return new GetHelper<Boolean>(catName, dbName, tableName, true, true) {
-      @Override
-      protected String describeResult() {
-        return "delete partition column stats";
-      }
-      @Override
-      protected Boolean getSqlResult(GetHelper<Boolean> ctx) throws MetaException {
-        DirectSqlDeleteStats deleteStats = new DirectSqlDeleteStats(directSql, pm);
-        return deleteStats.deletePartitionColumnStats(catName, dbName, tableName, partNames, cols, engine);
-      }
-      @Override
-      protected Boolean getJdoResult(GetHelper<Boolean> ctx)
-              throws MetaException, NoSuchObjectException, InvalidObjectException, InvalidInputException {
-        return deletePartitionColumnStatisticsViaJdo(catName, dbName, tableName, partNames, cols, engine);
-      }
-    }.run(false);
-  }
-
-  private boolean deletePartitionColumnStatisticsViaJdo(String catName, String dbName, String tableName,
-      List<String> partNames, List<String> colNames, String engine)
-      throws NoSuchObjectException, MetaException, InvalidObjectException, InvalidInputException {
-    boolean ret = false;
-    String database = org.apache.commons.lang3.StringUtils.defaultString(dbName,
-      Warehouse.DEFAULT_DATABASE_NAME);
-    String catalog = normalizeIdentifier(catName);
-    try {
-      openTransaction();
-      Batchable<String, Void> b = new Batchable<String, Void>() {
-        @Override
-        public List<Void> run(List<String> input) throws Exception {
-          Query query = pm.newQuery(MPartitionColumnStatistics.class);
-          addQueryAfterUse(query);
-          String filter;
-          String parameters;
-          if (colNames != null && !colNames.isEmpty()) {
-            filter = "t1.contains(partition.partitionName) && partition.table.database.name == t2 && partition.table.tableName == t3 && "
-                + "t4.contains(colName) && partition.table.database.catalogName == t5" + (engine != null ? " && engine == t6" : "");
-            parameters = "java.util.Collection t1, java.lang.String t2, java.lang.String t3, "
-                + "java.util.Collection t4, java.lang.String t5" + (engine != null ? ", java.lang.String t6" : "");
-          } else {
-            filter = "t1.contains(partition.partitionName) && partition.table.database.name == t2 && partition.table.tableName == t3 && " +
-                "partition.table.database.catalogName == t4" + (engine != null ? " && engine == t5" : "");
-            parameters = "java.util.Collection t1, java.lang.String t2, java.lang.String t3, java.lang.String t4" + (engine != null ? ", java.lang.String t5" : "");
-          }
-          query.setFilter(filter);
-          query.declareParameters(parameters);
-          List<Object> params = new ArrayList<>();
-          params.add(input);
-          params.add(normalizeIdentifier(database));
-          params.add(normalizeIdentifier(tableName));
-          if (colNames != null && !colNames.isEmpty()) {
-            params.add(colNames);
-          }
-          params.add(catalog);
-          if (engine != null) {
-            params.add(engine);
-          }
-          List<MPartitionColumnStatistics> mStatsObjColl =
-              (List<MPartitionColumnStatistics>) query.executeWithArray(params.toArray());
-          pm.retrieveAll(mStatsObjColl);
-          if (mStatsObjColl != null) {
-            pm.deletePersistentAll(mStatsObjColl);
-          }
-          return null;
-        }
-      };
-      try {
-        Batchable.runBatched(batchSize, partNames, b);
-      } finally {
-        b.closeAllQueries();
-      }
-
-      Batchable.runBatched(batchSize, partNames, new Batchable<String, Void>() {
-        @Override
-        public List<Void> run(List<String> input) throws MetaException {
-          Pair<Query, Map<String, String>> queryWithParams = getPartQueryWithParams(pm, catalog, database, tableName,
-              input);
-          try (QueryWrapper qw = new QueryWrapper(queryWithParams.getLeft())) {
-            qw.setResultClass(MPartition.class);
-            qw.setClass(MPartition.class);
-            List<MPartition> mparts = (List<MPartition>) qw.executeWithMap(queryWithParams.getRight());
-            for (MPartition mPart : mparts) {
-              Map<String, String> params = mPart.getParameters();
-              if (params != null && params.containsKey(StatsSetupConst.COLUMN_STATS_ACCURATE)) {
-                if (colNames == null || colNames.isEmpty()) {
-                  StatsSetupConst.clearColumnStatsState(params);
-                } else {
-                  StatsSetupConst.removeColumnStatsState(params, colNames);
-                }
-                mPart.setParameters(params);
-              }
-            }
-          }
-          return Collections.emptyList();
-        }
-      });
-      ret = commitTransaction();
-    } finally {
-      rollbackAndCleanup(ret, null);
-    }
-    return ret;
-  }
-
-  @Override
-  public boolean deleteTableColumnStatistics(String catName, String dbName, String tableName,
-      List<String> colNames, String engine)
-      throws NoSuchObjectException, MetaException, InvalidObjectException, InvalidInputException {
-    dbName = org.apache.commons.lang3.StringUtils.defaultString(dbName, Warehouse.DEFAULT_DATABASE_NAME);
-    if (tableName == null) {
-      throw new InvalidInputException("Table name is null.");
-    }
-    List<String> cols = normalizeIdentifiers(colNames);
-    return new GetHelper<Boolean>(catName, dbName, tableName, true, true) {
-      @Override
-      protected String describeResult() {
-        return "delete table column stats";
-      }
-      @Override
-      protected Boolean getSqlResult(GetHelper<Boolean> ctx) throws MetaException {
-        DirectSqlDeleteStats deleteStats = new DirectSqlDeleteStats(directSql, pm);
-        return deleteStats.deleteTableColumnStatistics(getTable(), cols, engine);
-      }
-      @Override
-      protected Boolean getJdoResult(GetHelper<Boolean> ctx)
-              throws MetaException, NoSuchObjectException, InvalidObjectException, InvalidInputException {
-        return deleteTableColumnStatisticsViaJdo(catName, dbName, tableName, cols, engine);
-      }
-    }.run(true);
-  }
-
-  private boolean deleteTableColumnStatisticsViaJdo(String catName, String dbName, String tableName,
-      List<String> colNames, String engine) throws NoSuchObjectException, MetaException, InvalidObjectException, InvalidInputException {
-    boolean ret = false;
-    Query query = null;
-    try {
-      openTransaction();
-      List<MTableColumnStatistics> mStatsObjColl;
-      // Note: this does not verify ACID state; called internally when removing cols/etc.
-      //       Also called via an unused metastore API that checks for ACID tables.
-      query = pm.newQuery(MTableColumnStatistics.class);
-      String filter;
-      String parameters;
-      if (colNames != null && !colNames.isEmpty()) {
-        filter = "table.tableName == t1 && table.database.name == t2 && table.database.catalogName == t3 && t4.contains(colName)" + (engine != null ? " && engine == t5" : "");
-        parameters = "java.lang.String t1, java.lang.String t2, java.lang.String t3, java.util.Collection t4" + (engine != null ? ", java.lang.String t5" : "");
-      } else {
-        filter = "table.tableName == t1 && table.database.name == t2 && table.database.catalogName == t3" + (engine != null ? " && engine == t4" : "");
-        parameters = "java.lang.String t1, java.lang.String t2, java.lang.String t3" + (engine != null ? ", java.lang.String t4" : "");
-      }
-
-      query.setFilter(filter);
-      query.declareParameters(parameters);
-      List<Object> params = new ArrayList<>();
-      params.add(normalizeIdentifier(tableName));
-      params.add(normalizeIdentifier(dbName));
-      params.add(catName == null ? null : normalizeIdentifier(catName));
-      if (colNames != null && !colNames.isEmpty()) {
-        params.add(colNames);
-      }
-      if (engine != null) {
-        params.add(engine);
-      }
-      mStatsObjColl = (List<MTableColumnStatistics>) query.executeWithArray(params.toArray());
-      pm.retrieveAll(mStatsObjColl);
-      if (mStatsObjColl != null) {
-        pm.deletePersistentAll(mStatsObjColl);
-      }
-
-      MTable mTable = getMTable(catName, dbName, tableName);
-      if (mTable != null) {
-        Map<String, String> tableParams = mTable.getParameters();
-        if (tableParams != null && tableParams.containsKey(StatsSetupConst.COLUMN_STATS_ACCURATE)) {
-          if (colNames == null || colNames.isEmpty()) {
-            StatsSetupConst.clearColumnStatsState(tableParams);
-          } else {
-            StatsSetupConst.removeColumnStatsState(tableParams, colNames);
-          }
-          mTable.setParameters(tableParams);
-        }
-      }
-      ret = commitTransaction();
-    } finally {
-      rollbackAndCleanup(ret, query);
-    }
-    return ret;
   }
 
   @Override
@@ -5251,583 +3199,6 @@ public class ObjectStore implements RawStore, Configurable {
       rollbackAndCleanup(committed, query);
     }
     return pkg;
-  }
-
-  private void executePlainSQL(String sql,
-      boolean atLeastOneRecord,
-      Consumer<Exception> exceptionConsumer)
-      throws SQLException, MetaException {
-    String s = dbType.getPrepareTxnStmt();
-    assert pm.currentTransaction().isActive();
-    JDOConnection jdoConn = pm.getDataStoreConnection();
-    Connection conn = (Connection) jdoConn.getNativeConnection();
-    try (Statement statement = conn.createStatement()) {
-      if (s != null) {
-        statement.execute(s);
-      }
-      try {
-        statement.execute(sql);
-        try (ResultSet rs = statement.getResultSet()) {
-          // sqlserver needs rs.next for validating the s4u nowait
-          if (atLeastOneRecord && !rs.next()) {
-            throw new MetaException("At least one record but none is returned from the query: " + sql);
-          }
-        }
-      } catch (SQLException e) {
-        if (exceptionConsumer != null) {
-          exceptionConsumer.accept(e);
-        } else {
-          throw e;
-        }
-      }
-    } finally {
-      jdoConn.close();
-    }
-  }
-
-  @Override
-  public List<SQLPrimaryKey> getPrimaryKeys(PrimaryKeysRequest request) throws MetaException {
-    try {
-      return getPrimaryKeysInternal(request.getCatName(),
-          request.getDb_name(),request.getTbl_name());
-    } catch (NoSuchObjectException e) {
-      throw new MetaException(ExceptionUtils.getStackTrace(e));
-    }
-  }
-
-  private List<SQLPrimaryKey> getPrimaryKeysInternal(final String catName,
-                                                     final String dbNameInput,
-                                                     final String tblNameInput)
-  throws MetaException, NoSuchObjectException {
-    final String dbName = dbNameInput != null ? normalizeIdentifier(dbNameInput) : null;
-    final String tblName = normalizeIdentifier(tblNameInput);
-    return new GetListHelper<SQLPrimaryKey>(catName, dbName, tblName, true, true) {
-
-      @Override
-      protected List<SQLPrimaryKey> getSqlResult(GetHelper<List<SQLPrimaryKey>> ctx) throws MetaException {
-        return directSql.getPrimaryKeys(catName, dbName, tblName);
-      }
-
-      @Override
-      protected List<SQLPrimaryKey> getJdoResult(
-        GetHelper<List<SQLPrimaryKey>> ctx) throws MetaException, NoSuchObjectException {
-        return getPrimaryKeysViaJdo(catName, dbName, tblName);
-      }
-    }.run(false);
-  }
-
-  private List<SQLPrimaryKey> getPrimaryKeysViaJdo(String catName, String dbName, String tblName) {
-    boolean commited = false;
-    List<SQLPrimaryKey> primaryKeys = null;
-    Query query = null;
-    try {
-      openTransaction();
-      query = pm.newQuery(MConstraint.class,
-        "parentTable.tableName == tbl_name && parentTable.database.name == db_name &&"
-        + " parentTable.database.catalogName == cat_name &&"
-        + " constraintType == MConstraint.PRIMARY_KEY_CONSTRAINT");
-      query.declareParameters("java.lang.String tbl_name, java.lang.String db_name, " +
-          "java.lang.String cat_name");
-      Collection<?> constraints = (Collection<?>) query.execute(tblName, dbName, catName);
-      pm.retrieveAll(constraints);
-      primaryKeys = new ArrayList<>();
-      for (Iterator<?> i = constraints.iterator(); i.hasNext();) {
-        MConstraint currPK = (MConstraint) i.next();
-        List<MFieldSchema> cols = currPK.getParentColumn() != null ?
-            currPK.getParentColumn().getCols() : currPK.getParentTable().getPartitionKeys();
-        int enableValidateRely = currPK.getEnableValidateRely();
-        boolean enable = (enableValidateRely & 4) != 0;
-        boolean validate = (enableValidateRely & 2) != 0;
-        boolean rely = (enableValidateRely & 1) != 0;
-        SQLPrimaryKey keyCol = new SQLPrimaryKey(dbName,
-            tblName,
-            cols.get(currPK.getParentIntegerIndex()).getName(),
-            currPK.getPosition(),
-            currPK.getConstraintName(), enable, validate, rely);
-        keyCol.setCatName(catName);
-        primaryKeys.add(keyCol);
-      }
-      commited = commitTransaction();
-    } finally {
-      rollbackAndCleanup(commited, query);
-    }
-    return primaryKeys;
-  }
-
-  private String getPrimaryKeyConstraintName(String catName, String dbName, String tblName) {
-    boolean commited = false;
-    String ret = null;
-    Query query = null;
-
-    try {
-      openTransaction();
-      query = pm.newQuery(MConstraint.class,
-        "parentTable.tableName == tbl_name && parentTable.database.name == db_name &&"
-        + " parentTable.database.catalogName == catName &&"
-        + " constraintType == MConstraint.PRIMARY_KEY_CONSTRAINT");
-      query.declareParameters("java.lang.String tbl_name, java.lang.String db_name, " +
-          "java.lang.String catName");
-      Collection<?> constraints = (Collection<?>) query.execute(tblName, dbName, catName);
-      pm.retrieveAll(constraints);
-      for (Iterator<?> i = constraints.iterator(); i.hasNext();) {
-        MConstraint currPK = (MConstraint) i.next();
-        ret = currPK.getConstraintName();
-        break;
-      }
-      commited = commitTransaction();
-     } finally {
-        rollbackAndCleanup(commited, query);
-     }
-     return ret;
-   }
-
-  @Override
-  public List<SQLForeignKey> getForeignKeys(ForeignKeysRequest request) throws MetaException {
-    try {
-      return getForeignKeysInternal(request.getCatName(),
-          request.getParent_db_name(), request.getParent_tbl_name() ,
-          request.getForeign_db_name(),request.getForeign_tbl_name(), true,
-          true);
-    } catch (NoSuchObjectException e) {
-      throw new MetaException(ExceptionUtils.getStackTrace(e));
-    }
-  }
-
-  private List<SQLForeignKey> getForeignKeysInternal(
-      final String catName, final String parent_db_name_input, final String parent_tbl_name_input,
-      final String foreign_db_name_input, final String foreign_tbl_name_input, boolean allowSql,
-      boolean allowJdo) throws MetaException, NoSuchObjectException {
-    final String parent_db_name = (parent_db_name_input != null) ? normalizeIdentifier(parent_db_name_input) : null;
-    final String parent_tbl_name = (parent_tbl_name_input != null) ? normalizeIdentifier(parent_tbl_name_input) : null;
-    final String foreign_db_name = (foreign_db_name_input != null) ? normalizeIdentifier(foreign_db_name_input) : null;
-    final String foreign_tbl_name = (foreign_tbl_name_input != null)
-                                    ? normalizeIdentifier(foreign_tbl_name_input) : null;
-    final String db_name;
-    final String tbl_name;
-    if (foreign_tbl_name == null) {
-      // The FK table name might be null if we are retrieving the constraint from the PK side
-      db_name = parent_db_name;
-      tbl_name = parent_tbl_name;
-    } else {
-      db_name = foreign_db_name;
-      tbl_name = foreign_tbl_name;
-    }
-    return new GetListHelper<SQLForeignKey>(catName, db_name, tbl_name, allowSql, allowJdo) {
-
-      @Override
-      protected List<SQLForeignKey> getSqlResult(GetHelper<List<SQLForeignKey>> ctx) throws MetaException {
-        return directSql.getForeignKeys(catName, parent_db_name,
-          parent_tbl_name, foreign_db_name, foreign_tbl_name);
-      }
-
-      @Override
-      protected List<SQLForeignKey> getJdoResult(
-        GetHelper<List<SQLForeignKey>> ctx) throws MetaException, NoSuchObjectException {
-        return getForeignKeysViaJdo(catName, parent_db_name,
-          parent_tbl_name, foreign_db_name, foreign_tbl_name);
-      }
-    }.run(false);
-  }
-
-  private List<SQLForeignKey> getForeignKeysViaJdo(String catName, String parentDbName,
-      String parentTblName, String foreignDbName, String foreignTblName) {
-    boolean commited = false;
-    List<SQLForeignKey> foreignKeys = null;
-    Collection<?> constraints = null;
-    Query query = null;
-    Map<String, String> tblToConstraint = new HashMap<>();
-    try {
-      openTransaction();
-      String queryText =
-          " parentTable.database.catalogName == catName1 &&" + "childTable.database.catalogName == catName2 && " + (
-              parentTblName != null ? "parentTable.tableName == parent_tbl_name && " : "") + (
-              parentDbName != null ? " parentTable.database.name == parent_db_name && " : "") + (
-              foreignTblName != null ? " childTable.tableName == foreign_tbl_name && " : "") + (
-              foreignDbName != null ? " childTable.database.name == foreign_db_name && " : "")
-              + " constraintType == MConstraint.FOREIGN_KEY_CONSTRAINT";
-      queryText = queryText.trim();
-      query = pm.newQuery(MConstraint.class, queryText);
-      String paramText = "java.lang.String catName1, java.lang.String catName2" + (
-          parentTblName == null ? "" : ", java.lang.String parent_tbl_name") + (
-          parentDbName == null ? "" : " , java.lang.String parent_db_name") + (
-          foreignTblName == null ? "" : ", java.lang.String foreign_tbl_name") + (
-          foreignDbName == null ? "" : " , java.lang.String foreign_db_name");
-      query.declareParameters(paramText);
-      List<String> params = new ArrayList<>();
-      params.add(catName);
-      params.add(catName); // This is not a mistake, catName is in the where clause twice
-      if (parentTblName != null) {
-        params.add(parentTblName);
-      }
-      if (parentDbName != null) {
-        params.add(parentDbName);
-      }
-      if (foreignTblName != null) {
-        params.add(foreignTblName);
-      }
-      if (foreignDbName != null) {
-        params.add(foreignDbName);
-      }
-      constraints = (Collection<?>) query.executeWithArray(params.toArray(new String[0]));
-
-      pm.retrieveAll(constraints);
-      foreignKeys = new ArrayList<>();
-      for (Iterator<?> i = constraints.iterator(); i.hasNext();) {
-        MConstraint currPKFK = (MConstraint) i.next();
-        List<MFieldSchema> parentCols = currPKFK.getParentColumn() != null ?
-            currPKFK.getParentColumn().getCols() : currPKFK.getParentTable().getPartitionKeys();
-        List<MFieldSchema> childCols = currPKFK.getChildColumn() != null ?
-            currPKFK.getChildColumn().getCols() : currPKFK.getChildTable().getPartitionKeys();
-        int enableValidateRely = currPKFK.getEnableValidateRely();
-        boolean enable = (enableValidateRely & 4) != 0;
-        boolean validate = (enableValidateRely & 2) != 0;
-        boolean rely = (enableValidateRely & 1) != 0;
-        String consolidatedtblName =
-          currPKFK.getParentTable().getDatabase().getName() + "." +
-          currPKFK.getParentTable().getTableName();
-        String pkName;
-        if (tblToConstraint.containsKey(consolidatedtblName)) {
-          pkName = tblToConstraint.get(consolidatedtblName);
-        } else {
-          pkName = getPrimaryKeyConstraintName(currPKFK.getParentTable().getDatabase().getCatalogName(),
-              currPKFK.getParentTable().getDatabase().getName(),
-              currPKFK.getParentTable().getTableName());
-          tblToConstraint.put(consolidatedtblName, pkName);
-        }
-        SQLForeignKey fk = new SQLForeignKey(
-          currPKFK.getParentTable().getDatabase().getName(),
-          currPKFK.getParentTable().getTableName(),
-          parentCols.get(currPKFK.getParentIntegerIndex()).getName(),
-          currPKFK.getChildTable().getDatabase().getName(),
-          currPKFK.getChildTable().getTableName(),
-          childCols.get(currPKFK.getChildIntegerIndex()).getName(),
-          currPKFK.getPosition(),
-          currPKFK.getUpdateRule(),
-          currPKFK.getDeleteRule(),
-          currPKFK.getConstraintName(), pkName, enable, validate, rely);
-        fk.setCatName(catName);
-        foreignKeys.add(fk);
-      }
-      commited = commitTransaction();
-    } finally {
-      rollbackAndCleanup(commited, query);
-    }
-    return foreignKeys;
-  }
-
-  @Override
-  public List<SQLUniqueConstraint> getUniqueConstraints(UniqueConstraintsRequest request) throws MetaException {
-    try {
-      return getUniqueConstraintsInternal(request.getCatName(),
-         request.getDb_name(),request.getTbl_name(), true, true);
-    } catch (NoSuchObjectException e) {
-      throw new MetaException(ExceptionUtils.getStackTrace(e));
-    }
-  }
-
-  private List<SQLUniqueConstraint> getUniqueConstraintsInternal(
-      String catNameInput, final String db_name_input, final String tbl_name_input,
-      boolean allowSql, boolean allowJdo) throws MetaException, NoSuchObjectException {
-    final String catName = normalizeIdentifier(catNameInput);
-    final String db_name = normalizeIdentifier(db_name_input);
-    final String tbl_name = normalizeIdentifier(tbl_name_input);
-    return new GetListHelper<SQLUniqueConstraint>(catName, db_name, tbl_name, allowSql, allowJdo) {
-
-      @Override
-      protected List<SQLUniqueConstraint> getSqlResult(GetHelper<List<SQLUniqueConstraint>> ctx)
-              throws MetaException {
-        return directSql.getUniqueConstraints(catName, db_name, tbl_name);
-      }
-
-      @Override
-      protected List<SQLUniqueConstraint> getJdoResult(GetHelper<List<SQLUniqueConstraint>> ctx)
-              throws MetaException, NoSuchObjectException {
-        return getUniqueConstraintsViaJdo(catName, db_name, tbl_name);
-      }
-    }.run(false);
-  }
-
-  private List<SQLUniqueConstraint> getUniqueConstraintsViaJdo(String catName, String dbName, String tblName) {
-    boolean commited = false;
-    List<SQLUniqueConstraint> uniqueConstraints = null;
-    Query query = null;
-    try {
-      openTransaction();
-      query = pm.newQuery(MConstraint.class,
-        "parentTable.tableName == tbl_name && parentTable.database.name == db_name && parentTable.database.catalogName == catName &&"
-        + " constraintType == MConstraint.UNIQUE_CONSTRAINT");
-      query.declareParameters("java.lang.String tbl_name, java.lang.String db_name, java.lang.String catName");
-      Collection<?> constraints = (Collection<?>) query.execute(tblName, dbName, catName);
-      pm.retrieveAll(constraints);
-      uniqueConstraints = new ArrayList<>();
-      for (Iterator<?> i = constraints.iterator(); i.hasNext();) {
-        MConstraint currConstraint = (MConstraint) i.next();
-        List<MFieldSchema> cols = currConstraint.getParentColumn() != null ?
-            currConstraint.getParentColumn().getCols() : currConstraint.getParentTable().getPartitionKeys();
-        int enableValidateRely = currConstraint.getEnableValidateRely();
-        boolean enable = (enableValidateRely & 4) != 0;
-        boolean validate = (enableValidateRely & 2) != 0;
-        boolean rely = (enableValidateRely & 1) != 0;
-        uniqueConstraints.add(new SQLUniqueConstraint(catName, dbName, tblName,
-            cols.get(currConstraint.getParentIntegerIndex()).getName(), currConstraint.getPosition(),
-            currConstraint.getConstraintName(), enable, validate, rely));
-      }
-      commited = commitTransaction();
-    } finally {
-      rollbackAndCleanup(commited, query);
-    }
-    return uniqueConstraints;
-  }
-
-  @Override
-  public List<SQLNotNullConstraint> getNotNullConstraints(NotNullConstraintsRequest request) throws MetaException {
-    try {
-      return getNotNullConstraintsInternal(request.getCatName(),request.getDb_name(),request.getTbl_name(), true, true);
-    } catch (NoSuchObjectException e) {
-      throw new MetaException(ExceptionUtils.getStackTrace(e));
-    }
-  }
-
-  @Override
-  public List<SQLDefaultConstraint> getDefaultConstraints(DefaultConstraintsRequest request) throws MetaException {
-    try {
-      return getDefaultConstraintsInternal(request.getCatName(),request.getDb_name(),request.getTbl_name(), true, true);
-    } catch (NoSuchObjectException e) {
-      throw new MetaException(ExceptionUtils.getStackTrace(e));
-    }
-  }
-
-  @Override
-  public List<SQLCheckConstraint> getCheckConstraints(CheckConstraintsRequest request) throws MetaException {
-    try {
-      return getCheckConstraintsInternal(request.getCatName(),request.getDb_name(),request.getTbl_name(), true, true);
-    } catch (NoSuchObjectException e) {
-      throw new MetaException(ExceptionUtils.getStackTrace(e));
-    }
-  }
-
-  private List<SQLDefaultConstraint> getDefaultConstraintsInternal(
-      String catName, final String db_name_input, final String tbl_name_input, boolean allowSql,
-      boolean allowJdo) throws MetaException, NoSuchObjectException {
-    catName = normalizeIdentifier(catName);
-    final String db_name = normalizeIdentifier(db_name_input);
-    final String tbl_name = normalizeIdentifier(tbl_name_input);
-    return new GetListHelper<SQLDefaultConstraint>(catName, db_name, tbl_name, allowSql, allowJdo) {
-
-      @Override
-      protected List<SQLDefaultConstraint> getSqlResult(GetHelper<List<SQLDefaultConstraint>> ctx)
-              throws MetaException {
-        return directSql.getDefaultConstraints(catName, db_name, tbl_name);
-      }
-
-      @Override
-      protected List<SQLDefaultConstraint> getJdoResult(GetHelper<List<SQLDefaultConstraint>> ctx)
-              throws MetaException, NoSuchObjectException {
-        return getDefaultConstraintsViaJdo(catName, db_name, tbl_name);
-      }
-    }.run(false);
-  }
-
-  protected List<SQLCheckConstraint> getCheckConstraintsInternal(String catName, final String db_name_input,
-                                                                     final String tbl_name_input, boolean allowSql,
-                                                                 boolean allowJdo)
-      throws MetaException, NoSuchObjectException {
-    final String db_name = normalizeIdentifier(db_name_input);
-    final String tbl_name = normalizeIdentifier(tbl_name_input);
-    return new GetListHelper<SQLCheckConstraint>(normalizeIdentifier(catName), db_name, tbl_name,
-        allowSql, allowJdo) {
-
-      @Override
-      protected List<SQLCheckConstraint> getSqlResult(GetHelper<List<SQLCheckConstraint>> ctx)
-          throws MetaException {
-        return directSql.getCheckConstraints(catName, db_name, tbl_name);
-      }
-
-      @Override
-      protected List<SQLCheckConstraint> getJdoResult(GetHelper<List<SQLCheckConstraint>> ctx)
-          throws MetaException, NoSuchObjectException {
-        return getCheckConstraintsViaJdo(catName, db_name, tbl_name);
-      }
-    }.run(false);
-  }
-
-  private List<SQLCheckConstraint> getCheckConstraintsViaJdo(String catName, String dbName, String tblName) {
-    boolean commited = false;
-    List<SQLCheckConstraint> checkConstraints= null;
-    Query query = null;
-    try {
-      openTransaction();
-      query = pm.newQuery(MConstraint.class,
-          "parentTable.tableName == tbl_name && parentTable.database.name == db_name &&"
-          + " parentTable.database.catalogName == catName && constraintType == MConstraint.CHECK_CONSTRAINT");
-      query.declareParameters("java.lang.String tbl_name, java.lang.String db_name, java.lang.String catName");
-      Collection<?> constraints = (Collection<?>) query.execute(tblName, dbName, catName);
-      pm.retrieveAll(constraints);
-      checkConstraints = new ArrayList<>();
-      for (Iterator<?> i = constraints.iterator(); i.hasNext();) {
-        MConstraint currConstraint = (MConstraint) i.next();
-        List<MFieldSchema> cols = currConstraint.getParentColumn() != null ?
-            currConstraint.getParentColumn().getCols() : currConstraint.getParentTable().getPartitionKeys();
-        int enableValidateRely = currConstraint.getEnableValidateRely();
-        boolean enable = (enableValidateRely & 4) != 0;
-        boolean validate = (enableValidateRely & 2) != 0;
-        boolean rely = (enableValidateRely & 1) != 0;
-        checkConstraints.add(new SQLCheckConstraint(catName, dbName, tblName,
-                                                        cols.get(currConstraint.getParentIntegerIndex()).getName(),
-                                                        currConstraint.getDefaultValue(),
-                                                    currConstraint.getConstraintName(), enable, validate, rely));
-      }
-      commited = commitTransaction();
-    } finally {
-      rollbackAndCleanup(commited, query);
-    }
-    return checkConstraints;
-  }
-
-  private List<SQLDefaultConstraint> getDefaultConstraintsViaJdo(String catName, String dbName, String tblName) {
-    boolean commited = false;
-    List<SQLDefaultConstraint> defaultConstraints= null;
-    Query query = null;
-    try {
-      openTransaction();
-      query = pm.newQuery(MConstraint.class,
-        "parentTable.tableName == tbl_name && parentTable.database.name == db_name &&"
-            + " parentTable.database.catalogName == catName &&"
-        + " constraintType == MConstraint.DEFAULT_CONSTRAINT");
-      query.declareParameters(
-          "java.lang.String tbl_name, java.lang.String db_name, java.lang.String catName");
-      Collection<?> constraints = (Collection<?>) query.execute(tblName, dbName, catName);
-      pm.retrieveAll(constraints);
-      defaultConstraints = new ArrayList<>();
-      for (Iterator<?> i = constraints.iterator(); i.hasNext();) {
-        MConstraint currConstraint = (MConstraint) i.next();
-        List<MFieldSchema> cols = currConstraint.getParentColumn() != null ?
-            currConstraint.getParentColumn().getCols() : currConstraint.getParentTable().getPartitionKeys();
-        int enableValidateRely = currConstraint.getEnableValidateRely();
-        boolean enable = (enableValidateRely & 4) != 0;
-        boolean validate = (enableValidateRely & 2) != 0;
-        boolean rely = (enableValidateRely & 1) != 0;
-        defaultConstraints.add(new SQLDefaultConstraint(catName, dbName, tblName,
-            cols.get(currConstraint.getParentIntegerIndex()).getName(), currConstraint.getDefaultValue(),
-            currConstraint.getConstraintName(), enable, validate, rely));
-      }
-      commited = commitTransaction();
-    } finally {
-      rollbackAndCleanup(commited, query);
-    }
-    return defaultConstraints;
-  }
-
-  protected List<SQLNotNullConstraint> getNotNullConstraintsInternal(String catName, final String db_name_input,
-                                                                     final String tbl_name_input, boolean allowSql, boolean allowJdo)
-      throws MetaException, NoSuchObjectException {
-    catName = normalizeIdentifier(catName);
-    final String db_name = normalizeIdentifier(db_name_input);
-    final String tbl_name = normalizeIdentifier(tbl_name_input);
-    return new GetListHelper<SQLNotNullConstraint>(catName, db_name, tbl_name, allowSql, allowJdo) {
-
-      @Override
-      protected List<SQLNotNullConstraint> getSqlResult(GetHelper<List<SQLNotNullConstraint>> ctx)
-          throws MetaException {
-        return directSql.getNotNullConstraints(catName, db_name, tbl_name);
-      }
-
-      @Override
-      protected List<SQLNotNullConstraint> getJdoResult(GetHelper<List<SQLNotNullConstraint>> ctx)
-          throws MetaException, NoSuchObjectException {
-        return getNotNullConstraintsViaJdo(catName, db_name, tbl_name);
-      }
-    }.run(false);
-  }
-
-  private List<SQLNotNullConstraint> getNotNullConstraintsViaJdo(String catName, String dbName, String tblName) {
-    boolean commited = false;
-    List<SQLNotNullConstraint> notNullConstraints = null;
-    Query query = null;
-    try {
-      openTransaction();
-      query = pm.newQuery(MConstraint.class,
-          "parentTable.tableName == tbl_name && parentTable.database.name == db_name &&"
-              + " parentTable.database.catalogName == catName && constraintType == MConstraint.NOT_NULL_CONSTRAINT");
-      query.declareParameters(
-          "java.lang.String tbl_name, java.lang.String db_name, java.lang.String catName");
-      Collection<?> constraints = (Collection<?>) query.execute(tblName, dbName, catName);
-      pm.retrieveAll(constraints);
-      notNullConstraints = new ArrayList<>();
-      for (Iterator<?> i = constraints.iterator(); i.hasNext();) {
-        MConstraint currConstraint = (MConstraint) i.next();
-        List<MFieldSchema> cols = currConstraint.getParentColumn() != null ?
-            currConstraint.getParentColumn().getCols() : currConstraint.getParentTable().getPartitionKeys();
-        int enableValidateRely = currConstraint.getEnableValidateRely();
-        boolean enable = (enableValidateRely & 4) != 0;
-        boolean validate = (enableValidateRely & 2) != 0;
-        boolean rely = (enableValidateRely & 1) != 0;
-        notNullConstraints.add(new SQLNotNullConstraint(catName, dbName,
-            tblName,
-            cols.get(currConstraint.getParentIntegerIndex()).getName(),
-            currConstraint.getConstraintName(), enable, validate, rely));
-      }
-      commited = commitTransaction();
-    } finally {
-      rollbackAndCleanup(commited, query);
-    }
-    return notNullConstraints;
-  }
-
-  /**
-   * Api to fetch all constraints at once
-   * @param request request object
-   * @return all table constraints
-   * @throws MetaException
-   * @throws NoSuchObjectException
-   */
-  @Override
-  public SQLAllTableConstraints getAllTableConstraints(AllTableConstraintsRequest request)
-      throws MetaException, NoSuchObjectException {
-    String catName = request.getCatName();
-    String dbName = request.getDbName();
-    String tblName = request.getTblName();
-    debugLog("Get all table constraints for the table - " + catName + "." + dbName + "." + tblName
-        + " in class ObjectStore.java");
-    SQLAllTableConstraints sqlAllTableConstraints = new SQLAllTableConstraints();
-    PrimaryKeysRequest primaryKeysRequest = new PrimaryKeysRequest(dbName, tblName);
-    primaryKeysRequest.setCatName(catName);
-    sqlAllTableConstraints.setPrimaryKeys(getPrimaryKeys(primaryKeysRequest));
-    ForeignKeysRequest foreignKeysRequest =
-        new ForeignKeysRequest(null, null, dbName, tblName);
-    foreignKeysRequest.setCatName(catName);
-    sqlAllTableConstraints.setForeignKeys(getForeignKeys(foreignKeysRequest));
-    sqlAllTableConstraints.
-        setUniqueConstraints(getUniqueConstraints(new UniqueConstraintsRequest(catName, dbName, tblName)));
-    sqlAllTableConstraints.
-        setDefaultConstraints(getDefaultConstraints(new DefaultConstraintsRequest(catName, dbName, tblName)));
-    sqlAllTableConstraints.
-        setCheckConstraints(getCheckConstraints(new CheckConstraintsRequest(catName, dbName, tblName)));
-    sqlAllTableConstraints.
-        setNotNullConstraints(getNotNullConstraints(new NotNullConstraintsRequest(catName, dbName, tblName)));
-    return sqlAllTableConstraints;
-  }
-
-  @Override
-  public void dropConstraint(String catName, String dbName, String tableName,
-                             String constraintName, boolean missingOk)
-      throws NoSuchObjectException {
-    boolean success = false;
-    try {
-      openTransaction();
-
-      List<MConstraint> tabConstraints =
-          listAllTableConstraintsWithOptionalConstraintName(catName,  dbName, tableName, constraintName);
-      if (CollectionUtils.isNotEmpty(tabConstraints)) {
-        pm.deletePersistentAll(tabConstraints);
-      } else if (!missingOk) {
-        throw new NoSuchObjectException("The constraint: " + constraintName +
-          " does not exist for the associated table: " + dbName + "." + tableName);
-      }
-      success = commitTransaction();
-    } finally {
-      rollbackAndCleanup(success, null);
-    }
   }
 
   @Override
@@ -7463,72 +4834,6 @@ public class ObjectStore implements RawStore, Configurable {
       }
       return ret;
     }
-  }
-
-  /**
-   * Return true if the current statistics in the Metastore is valid
-   * for the query of the given "txnId" and "queryValidWriteIdList".
-   *
-   * Note that a statistics entity is valid iff
-   * the stats is written by the current query or
-   * the conjunction of the following two are true:
-   * ~ COLUMN_STATE_ACCURATE(CSA) state is true
-   * ~ Isolation-level (snapshot) compliant with the query
-   * @param tbl                    MTable of the stats entity
-   * @param queryValidWriteIdList  valid writeId list of the query
-   * @Precondition   "tbl" should be retrieved from the TBLS table.
-   */
-  private boolean isCurrentStatsValidForTheQuery(MTable tbl, String queryValidWriteIdList,
-      boolean isCompleteStatsWriter) throws MetaException {
-    return isCurrentStatsValidForTheQuery(tbl.getParameters(), tbl.getWriteId(),
-        queryValidWriteIdList, isCompleteStatsWriter);
-  }
-
-  /**
-   * Return true if the current statistics in the Metastore is valid
-   * for the query of the given "txnId" and "queryValidWriteIdList".
-   *
-   * Note that a statistics entity is valid iff
-   * the stats is written by the current query or
-   * the conjunction of the following two are true:
-   * ~ COLUMN_STATE_ACCURATE(CSA) state is true
-   * ~ Isolation-level (snapshot) compliant with the query
-   * @param queryValidWriteIdList  valid writeId list of the query
-   */
-  // TODO: move to somewhere else
-  public static boolean isCurrentStatsValidForTheQuery(
-      Map<String, String> statsParams, long statsWriteId, String queryValidWriteIdList,
-      boolean isCompleteStatsWriter) throws MetaException {
-
-    // Note: can be changed to debug/info to verify the calls.
-    LOG.debug("isCurrentStatsValidForTheQuery with stats write ID {}; query {}; writer: {} params {}",
-        statsWriteId, queryValidWriteIdList, isCompleteStatsWriter, statsParams);
-    // return true since the stats does not seem to be transactional.
-    if (statsWriteId < 1) {
-      return true;
-    }
-    // This COLUMN_STATS_ACCURATE(CSA) state checking also includes the case that the stats is
-    // written by an aborted transaction but TXNS has no entry for the transaction
-    // after compaction. Don't check for a complete stats writer - it may replace invalid stats.
-    if (!isCompleteStatsWriter && !StatsSetupConst.areBasicStatsUptoDate(statsParams)) {
-      return false;
-    }
-
-    if (queryValidWriteIdList != null) { // Can be null when stats are being reset to invalid.
-      ValidWriteIdList list4TheQuery = ValidReaderWriteIdList.fromValue(queryValidWriteIdList);
-      // Just check if the write ID is valid. If it's valid (i.e. we are allowed to see it),
-      // that means it cannot possibly be a concurrent write. If it's not valid (we are not
-      // allowed to see it), that means it's either concurrent or aborted, same thing for us.
-      if (list4TheQuery.isWriteIdValid(statsWriteId)) {
-        return true;
-      }
-      // Updater is also allowed to overwrite stats from aborted txns, as long as they are not concurrent.
-      if (isCompleteStatsWriter && list4TheQuery.isWriteIdAborted(statsWriteId)) {
-        return true;
-      }
-    }
-
-    return false;
   }
 
   @Override
