+++ standalone-metastore/metastore-server/src/main/java/org/apache/hadoop/hive/metastore/metastore/iface/ConstraintStore.java
@@ -1,0 +1,8 @@
+package org.apache.hadoop.hive.metastore.metastore.iface;
+
+import org.apache.hadoop.hive.metastore.metastore.MetaDescriptor;
+import org.apache.hadoop.hive.metastore.metastore.impl.ConstraintStoreImpl;
+
+@MetaDescriptor(alias = "constraint", defaultImpl = ConstraintStoreImpl.class)
+public interface ConstraintStore {
+}
